Abstract Array Subtypes

General Storage Arrays

TransferFunctions.OuterInnerArray — Type

OuterInnerArray{T,N,M,K,IA,OA} <: AbstractArray{T,N}

N-dimensional array that consists of an M dimensional 'outer' array of type OA<:AbstractArray{IA,M} of K dimensional T valued 'inner' arrays of type IA{T,K}. The constructor ensures that M+K==N and that the inner arrays have the same axes.

See also Slices

Fields

outer::OA

isinnerdim::SVector{N,Bool}

size::Size{N}

axes::Indices{N}

A container array that is used in the implementation of other array types that holds some dimensions in an inner array and these

Filtering Arrays

TransferFunctions.CirculantTensor — Type

CirculantTensor{T,N,M,AA} <: AbstractArray{T,N}

N-dimensional circulant tensor of an M dimensional array A or type AA<:AbstractArray{T,M}. The dimensionality N is equal to 2M, where the first M dimensions have interior filtering coordinates ct.interior given by the kernel axes ct.kern and the axes of A and the tail M dimensions have the kernel coordinates ct.kern.

A correlation filtering result of A with a kernel array Kcan be obtained by outer tensor contraction over the tail M dimensions of the CirculantTensor(A,K).

See also FilteringMatrix, imfilter

TransferFunctions.FilteringMatrix — Type

FilteringMatrix{T,M,P} <: AbstractMatrix{T}

A matrix that for a given kernel K, array A of size size(A)==size(K) equal to M and a padding scheme P gives an array F such that filtering (correlation or convolution if reflect(K) is used as the kernel) output of the kernel K and an array as A can be computed as F * A[:]. It is a view into the CirculantTensor for the kernel array K.

Data Arrays

For representing data that is sampled at some constant rate, possibly different in every dimension, there are SampledArrays that store the data coupled with its sampling.

TransferFunctions.SampledArray — Type

SampledArray{T,ST,N,AA<:AbstractArray} <: AbstractArray{T,N}

An N-dimensional sampled array with data of type T sample spacing of type ST.

See also SampledMatrix, SampledVector, SpatialArray

TransferFunctions.SampledMatrix — Type

SampledMatrix{T, ST} <: AbstractMatrix{T}

Two-dimensional array with elements of type T with a given sample spacing of type ST. Alias for SampledArray{T,ST,2}.

TransferFunctions.SampledVector — Type

SampledVector{T, ST} <: AbstractVector{T}

One-dimensional array with elements of type T with a given sample spacing of type ST. Alias for SampledArray{T,ST,1}.

Since sampling at spatially defined lattice points is so common in the context of microscopy, constants for SampledArrays with spatial sampling are provided as aliases.

TransferFunctions.SpatialArray — Type

SpatialArray{T,N} <: AbstractArray{T,N}

An N-dimensional array representing values sampled uniformly in space. Alias for SampledArray{T,<:Length,N}.

See also SpatialMatrix, SpatialVector, SampledArray

TransferFunctions.SpatialMatrix — Type

SpatialMatrix{T} <: AbstractMatrix{T}

Two-dimensional array with elements of type T with a given sampling distance in both directions. Alias for SpatialArray{T,2}.

TransferFunctions.SpatialVector — Type

SpatialVector{T} <: AbstractVector{T}

One-dimensional array with elements of type T with a given sampling distance. Alias for SpatialArray{T,1}.