Unit Types

TransferFunctions.Frequency — Type

TransferFunctions.Frequency{T, U}

A supertype for quantities and levels of dimension Unitful.𝐋 ^ -1 with a value of type T and units U.

See also: Unitful.Quantity, Unitful.Level.

TransferFunctions.FrequencyUnits — Type

TransferFunctions.FrequencyUnits{U}

A supertype for units of dimension Unitful.𝐋 ^ -1. Equivalent to Unitful.Units{U, Unitful.𝐋 ^ -1}.

See also: Unitful.Units.

TransferFunctions.FrequencyFreeUnits — Type

TransferFunctions.FrequencyFreeUnits{U}

A supertype for Unitful.FreeUnits of dimension Unitful.𝐋 ^ -1. Equivalent to Unitful.FreeUnits{U, Unitful.𝐋 ^ -1}.

Transfer Functions

TransferFunctions.TransferFunction — Type

TransferFunction

Super type for all transfer functions

A transfer function in microscopy is an object specifying the response in the image plane to the light coming from the object plane. There are several characteristics of a transfer functions which can influence the character of the response. Most commonly we are concerned with its spatial variance/invariance, i.e. whether the response is dependent of the objects position in the object plane, and linearity (which is almost always satisfied).

See also LinearShiftInvariantTransferFunction, ImpulseResponseMapping

Linear Transfer Functions

TransferFunctions.LinearShiftInvariantTransferFunction — Type

LinearShiftInvariantTransferFunction <: TransferFunction

A supertype for all linear transfer functions.

A linear transfer function is that which ensures the system to have a linear response. This means that the system obeys the superposition principle, i.e. the response to a superposition of inputs is a superposition of the corresponding responses, and shift-invariance principle, i.e. the response to a signal is invariant to translation, which can be restated as "The single object in the object plane produces the same output in the image plane irrespective of its position within the object plane."

See also TransferFunctions

TransferFunctions.conv — Function

conv(ltf::LinearShiftInvariantTransferFunction, img::SpatialArray{<:Real,2})

Transfer the image img using the linear transfer function ltf, i.e. convolve the image with the equivalent PSF.

conv(tf::PointSpreadFunction, img::SpatialArray{<:Real,2}, [args...]; <kwargs>)

Convolve the image img with the PSF tf. Additional arguments are passed to imfilter.

Keyword arguments

border=nothing If the border is a NamedTuple with the keys border of a type compatible with

BorderArray and apodization of type Apodization.ApodizationFunction, the border is applied to the image with the size of the FWHM of the PSF in the corresponding directions with the border and apodization used for edge tapering (taperedges)) and the full expanded array is returned.

TransferFunctions.deconv — Function

deconv(ltf::LinearShiftInvariantTransferFunction, img::SpatialArray{<:Real,2})

Deconvolve the image img that was transferred using the linear transfer function ltf using the specified algorithm.

Optical Transfer Functions

TransferFunctions.OpticalTransferFunction — Type

OpticalTransferFunction <: LinearShiftInvariantTransferFunction

Implementation

To create a new Optical transfer function (OTF) A <: OpticalTransferFunction, you must define the attenuation at a given frequency coordinate attenuation(otf::A, kx::Frequency, ky::Frequency).

TransferFunctions.OTFArray — Type

OTFArray{T<:Number} <: OpticalTransferFunction

Fields

data::AbstractMatrix{T}

Δxy::PixelSize{2}

center::Coordinate{2}

extend::ExtensionMethod

TransferFunctions.RadialOTF — Type

RadialOTF <: OpticalTransferFunction

If the pupil function of the system is symmetric, the OTF as well as the PSF are radially symmetric which can be used to optimize the calculations.

Implementation

attenuation(model::A, fr::Frequency)

cutoff(model::A, [a=0]) returning the largest frequency f such that attenuation(model, f) >= a if a > 0 and

attenuation(model, f) > 0 if a = 0.

TransferFunctions.CircularPupilOTF — Type

CircularPupilOTF(ρ₀::Real)

Ideal (aberration free) OTF of a diffraction limited imaging system with incoherent light with the cutoff-frequency ρ₀

TransferFunctions.otf — Function

otf(tf::OpticalTransferFunction, Δxy::PixelSize{2}, wh::Dims{2})

TransferFunctions.insupport — Function

insupport(otf::OpticalTransferFunction, fx::Frequency, fy::Frequency)

TransferFunctions.cutoff — Function

cutoff(::RadialOTF, a=0.0)

TransferFunctions.attenuation — Function

attenuation(otf::OpticalTransferFunction, ::Frequency, ::Frequency)

Point Spread Functions

TransferFunctions.PointSpreadFunction — Type

PointSpreadFunction <: LinearShiftInvariantTransferFunction

A point spread function is a description of a transfer functions specifying the intensity transfer of a single point source in the object plane into a region of the image plane.

Implementation

To create a new Point spread function (PSF) A <: PointSpreadFunction, you must define the intensity at a given length coordinate intensity(psf::A, x::Length, y::Length).

TransferFunctions.MeasuredPSF — Type

MeasuredPSF <: PointSpreadFunction

TransferFunctions.PSFArray — Type

PSFArray{T<:Real} <: PointSpreadFunction

A measurement of a PSF from the acquired image(s).

Fields

data::AbstractMatrix{T}

Δ::PixelSize{2}

center::Coordinate{2}

extend::ExtensionMethod

TransferFunctions.ModelPSF — Type

ModelPSF <: PointSpreadFunction

TransferFunctions.RadialPSF — Type

RadialPSF <: ModelPSF

TransferFunctions.BornWolf — Type

BornWolf{T<:Real} <: RadialPSF

Born & Wolf model point spread function

Fields

+ `λ::Length{T}`: emission wavelength 
+ `n::T`: index of refraction of the immersion medium
+ `NA::T`: numerical aperture of the objective

TransferFunctions.GibsonLanni — Type

GibsonLanni{T<:Real} <: RadialPSF

Gibson & Lanni model point spread function

Fields

+ `λ::Length{T}`: emission wavelength
+ `NA::T`: numerical aperture of the objective
+ `n_i::T`: refractive index of the immersion medium
+ `n_s::T`: refractive index of the sample
+ `n_g::T`: refractive index of the coverslip
+ `t_i::Length{T}`: working distance of the objective
+ `t_g::Length{T}`: coverslip thickness

TransferFunctions.psf — Function

psf(tf::PointSpreadFunction, Δ::PixelSize{2}, wh::Dims{2})

Generate a PSF array size wh for the model tf with the pixel size Δ.

TransferFunctions.intensity — Function

intensity(psf::PointSpreadFunction, x::Length, y::Length)

The intensity of a PSF at a given location.

TransferFunctions.fit — Function

fit(::ModelPSF, ::SpatialArray)

TransferFunctions.FWHM — Function

FWHM(psf::PointSpreadFunction)

Find the FWHM of the PSF psf in the x and y directions.

julia> tf = TransferFunctions.BornWolf(λ=488u"nm", NA=1.4)
BornWolf{Float64}(488.0 nm, 1.4, 1.3333333333333333)

julia> TransferFunctions.FWHM(tf)
((-119.55929936703521 nm, 119.55929936703521 nm), (-119.55929936703521 nm, 119.55929936703521 nm))

Non-linear Transfer Functions

TransferFunctions.NonLinearTransferFunction — Type

NonLinearShiftInvariantTransferFunction <: TransferFunction

A supertype for all non-linear transfer functions.

See also TransferFunctions, LinearShiftInvariantTransferFunction

TransferFunctions.ImpulseResponseMapping — Type

ImpulseResponseMapping <: NonLinearTransferFunction

An impulse response mapping is a description of a transfer function that prescribes an impulse response to every point in the object plane. This may be a point spread function or some other function that prescribes the response to the object plane point in the image plane.

An transfer defined by an impulse response mapping is a linear system, but is not translation invariant hence does not belong to the class of linear transfer functions.

Internals

Functions

TransferFunctions.roundcenter — Function

roundcenter(r::RoundingMode, A)

Calculate the center index of A (::AbstractArray/::Indices/::Size/::AbstractUnitRange) rounded with the mode r::RoundingMode.

See also roundupcenter, rounddowncenter, exactcenter

TransferFunctions.roundupcenter — Function

roundupcenter(A)

Calculate center index of A rounded-up (i.e. fft center).

Examples

julia> TF.roundupcenter(ones(15,15))
CartesianIndex(8, 8)

julia> TF.roundupcenter((16,16))
CartesianIndex(9, 9)

TransferFunctions.rounddowncenter — Function

rounddowncenter(A)

Calculate center index of A rounded-down (i.e. ifft center).

Examples

julia> TF.rounddowncenter(ones(15,15))
CartesianIndex(8, 8)

julia> TF.rounddowncenter((16,16))
CartesianIndex(8, 8)

TransferFunctions.exactcenter — Function

 exactcenter(A)

Calculate the exact center coordinate of A.

Examples

julia> TF.exactcenter(ones(15,15))
(8.0, 8.0)

julia> TF.exactcenter((16,16))
(8.5, 8.5)

TransferFunctions.aroundorigin — Function

aroundorigin(s::Size, origin=(0,0,...))

Return the offset axes of an array of size s around the origin.

TransferFunctions.contained — Function

contained(A, loc::Coordinate{N})

Return true if the coordinate loc is contained in the axes of array A.

TransferFunctions.interior — Function

interior(inds::Indices{N}, kern::Indices{N})

Return 'valid' indices for convolution of an array with indices inds with a kernel having the indices kern.

TransferFunctions.fftfreqs — Function

fftfreqs(sz::Size{2}, Δ::PixelSize{2})
fftfreqs(sz::Size{2}, Δ::Length)

Generate a frequency grid for 2D FFTs with the pixel size Δ

julia> x_f, y_f = TF.fftfreqs((5,5), 50u"nm");

julia> x_f
5×5 Matrix{Quantity{Float64, 𝐋^-1, Unitful.FreeUnits{(nm^-1,), 𝐋^-1, nothing}}}:
  0.0 nm^-1     0.0 nm^-1     0.0 nm^-1     0.0 nm^-1     0.0 nm^-1
  0.004 nm^-1   0.004 nm^-1   0.004 nm^-1   0.004 nm^-1   0.004 nm^-1
  0.008 nm^-1   0.008 nm^-1   0.008 nm^-1   0.008 nm^-1   0.008 nm^-1
 -0.008 nm^-1  -0.008 nm^-1  -0.008 nm^-1  -0.008 nm^-1  -0.008 nm^-1
 -0.004 nm^-1  -0.004 nm^-1  -0.004 nm^-1  -0.004 nm^-1  -0.004 nm^-1

julia> y_f
5×5 Matrix{Quantity{Float64, 𝐋^-1, Unitful.FreeUnits{(nm^-1,), 𝐋^-1, nothing}}}:
 0.0 nm^-1  0.004 nm^-1  0.008 nm^-1  -0.008 nm^-1  -0.004 nm^-1
 0.0 nm^-1  0.004 nm^-1  0.008 nm^-1  -0.008 nm^-1  -0.004 nm^-1
 0.0 nm^-1  0.004 nm^-1  0.008 nm^-1  -0.008 nm^-1  -0.004 nm^-1
 0.0 nm^-1  0.004 nm^-1  0.008 nm^-1  -0.008 nm^-1  -0.004 nm^-1
 0.0 nm^-1  0.004 nm^-1  0.008 nm^-1  -0.008 nm^-1  -0.004 nm^-1

TransferFunctions.posgrid — Function

posgrid(a::SampledArray)

Returns the grid of positions of the samples of a.

julia> sa = SampledArray(reshape(1:16, (4,4)), 61u"nm");

julia> xs, ys = TF.posgrid(sa);

julia> xs 
4×4 Matrix{Quantity{Float64, 𝐋, Unitful.FreeUnits{(nm,), 𝐋, nothing}}}:
   0.0 nm    0.0 nm    0.0 nm    0.0 nm
  61.0 nm   61.0 nm   61.0 nm   61.0 nm
 122.0 nm  122.0 nm  122.0 nm  122.0 nm
 183.0 nm  183.0 nm  183.0 nm  183.0 nm

julia> ys 
4×4 Matrix{Quantity{Float64, 𝐋, Unitful.FreeUnits{(nm,), 𝐋, nothing}}}:
 0.0 nm  61.0 nm  122.0 nm  183.0 nm
 0.0 nm  61.0 nm  122.0 nm  183.0 nm
 0.0 nm  61.0 nm  122.0 nm  183.0 nm
 0.0 nm  61.0 nm  122.0 nm  183.0 nm

posgrid(sz::Size{2}, Δ::PixelSize{2}; center=roundupcenter(sz))

Generate a position grid for 2D sampled images with the pixel size Δ

julia> x_p, y_p = TF.posgrid((5, 5), 50u"nm");

julia> x_p
5×5 Matrix{Quantity{Float64, 𝐋, Unitful.FreeUnits{(nm,), 𝐋, nothing}}}:
 -100.0 nm  -100.0 nm  -100.0 nm  -100.0 nm  -100.0 nm
  -50.0 nm   -50.0 nm   -50.0 nm   -50.0 nm   -50.0 nm
    0.0 nm     0.0 nm     0.0 nm     0.0 nm     0.0 nm
   50.0 nm    50.0 nm    50.0 nm    50.0 nm    50.0 nm
  100.0 nm   100.0 nm   100.0 nm   100.0 nm   100.0 nm

julia> y_p
5×5 Matrix{Quantity{Float64, 𝐋, Unitful.FreeUnits{(nm,), 𝐋, nothing}}}:
 -100.0 nm  -50.0 nm  0.0 nm  50.0 nm  100.0 nm
 -100.0 nm  -50.0 nm  0.0 nm  50.0 nm  100.0 nm
 -100.0 nm  -50.0 nm  0.0 nm  50.0 nm  100.0 nm
 -100.0 nm  -50.0 nm  0.0 nm  50.0 nm  100.0 nm
 -100.0 nm  -50.0 nm  0.0 nm  50.0 nm  100.0 nm

TransferFunctions.fillsize — Function

fillsize(Δ::Length, N::Integer) => PixelSize{N}

Construct a PixelSize with the same sampling Δ in every direction.

Type Aliases

TransferFunctions.Size — Type

Size{N}

N-dimensional size of an array. Alias for NTuple{N,Int}

TransferFunctions.Coordinate — Type

Coordinate{N,T}

N-dimensional coordinate. Alias for NTuple{N,T<:Real}

TransferFunctions.PixelSize — Type

PixelSize{N}

N-dimensional pixel size. Alias for NTuple{N,Length}

To Sort

TransferFunctions.IndependentIndexMapBorder — Type

IndependentIndexMapBorder{T} <: IndexMapBorder{T}

A border that maps the indices of the border independently for each dimension

TransferFunctions.IndexMapBorder — Type

IndexMapBorder{T} <: AbstractBorder{T}

A border that maps its values the values of the parent array with a certain index mapping

TransferFunctions.Reflect — Type

Reflect{T} <: IndexMapBorder{T}

A border that reflects the values of the border around the edges of the parent array

julia> TransferFunctions.BorderArray(reshape(1:16, (4,4)), TransferFunctions.Reflect, 2)
8×8 border_array(reshape(::UnitRange{Int64}, 4, 4), :Reflect) with eltype Int64 with indices -1:6×-1:6:
 6  2  2  6  10  14  14  10
 5  1  1  5   9  13  13   9
 5  1  1  5   9  13  13   9
 6  2  2  6  10  14  14  10
 7  3  3  7  11  15  15  11
 8  4  4  8  12  16  16  12
 8  4  4  8  12  16  16  12
 7  3  3  7  11  15  15  11

TransferFunctions.kern_padding — Function

kern_padding(kern)

Determine the padding necessary to keep the input array fully contained in the interior of the output when filtered with kern.

TransferFunctions.Fill — Type

Fill{T} <: AbstractBorder{T}

A border that fills the values with value::T

TransferFunctions.ind2sub — Function

ind2sub(inds::DimsInteger, ind::Integer)

Convert a linear index ind to a subscript index for the array size inds.

TransferFunctions.Circular — Type

Circular{T} <: IndexMapBorder{T}

A border that wraps the values around the edges of the parent array

julia> TransferFunctions.BorderArray(reshape(1:16, (4,4)), TransferFunctions.Circular, 2)
8×8 border_array(reshape(::UnitRange{Int64}, 4, 4), :Circular) with eltype Int64 with indices -1:6×-1:6:
 11  15  3  7  11  15  3  7
 12  16  4  8  12  16  4  8
  9  13  1  5   9  13  1  5
 10  14  2  6  10  14  2  6
 11  15  3  7  11  15  3  7
 12  16  4  8  12  16  4  8
  9  13  1  5   9  13  1  5
 10  14  2  6  10  14  2  6

TransferFunctions.InvalidBorderExtent — Type

InvalidBorderExtent <: Exception

Is thrown when trying to construct a BorderArray with padding larger than the maximum padding that would produce valid values.

TransferFunctions.Replicate — Type

Replicate{T}  <: IndexMapBorder{T}

A border that replicates the closest edge value of the parent array

julia> TransferFunctions.BorderArray(reshape(1:16, (4,4)), TransferFunctions.Replicate, 2)
8×8 border_array(reshape(::UnitRange{Int64}, 4, 4), :Replicate) with eltype Int64 with indices -1:6×-1:6:
 1  1  1  5   9  13  13  13
 1  1  1  5   9  13  13  13
 1  1  1  5   9  13  13  13
 2  2  2  6  10  14  14  14
 3  3  3  7  11  15  15  15
 4  4  4  8  12  16  16  16
 4  4  4  8  12  16  16  16
 4  4  4  8  12  16  16  16

TransferFunctions.Symmetric — Type

Symmetric{T} <: IndexMapBorder{T}

A border that symmetrically continues the values or the parent array around the edges, i.e. mirrors around the last valid parent index.

julia> TransferFunctions.BorderArray(reshape(1:16, (4,4)), TransferFunctions.Symmetric, 2)
8×8 border_array(reshape(::UnitRange{Int64}, 4, 4), :Symmetric) with eltype Int64 with indices -1:6×-1:6:
 11  7  3  7  11  15  11  7
 10  6  2  6  10  14  10  6
  9  5  1  5   9  13   9  5
 10  6  2  6  10  14  10  6
 11  7  3  7  11  15  11  7
 12  8  4  8  12  16  12  8
 11  7  3  7  11  15  11  7
 10  6  2  6  10  14  10  6