Filtering Arrays and Circulant Tensors

TransferFunctions.kern_padding is used to determine the padding necessary to perform the full domain filtering of an array.

TransferFunctions.kern_padding — Function

kern_padding(kern)

Determine the padding necessary to keep the input array fully contained in the interior of the output when filtered with kern.

TransferFunctions.ind2sub is used in the FilteringMatrix to determine the "kernel" and "parent" indices from the index to the FilteringMatrix.

TransferFunctions.ind2sub — Function

ind2sub(inds::DimsInteger, ind::Integer)

Convert a linear index ind to a subscript index for the array size inds.

Border Arrays

The error TransferFunctions.InvalidBorderExtent is used when the border type does not support the extent requested in the BorderArray constructor. If you implement your border type, you must implement TransferFunctions.validextension

TransferFunctions.validextension — Function

validextension(b::AbstractBorder, A::AbstractArray)

Returns the maximum padding extension that is valid for the border b and the parent array A.

julia> TF.validextension(TF.Circular(), ones(40,40))
((40, 40), (40, 40))

julia> TF.validextension(TF.Symmetric(), ones(40,40))
((39, 39), (39, 39))

julia> TF.validextension(TF.Fill(0), ones(40,40))
((9223372036854775807, 9223372036854775807), (9223372036854775807, 9223372036854775807))

If the border type that you want to implement is determined by an index map on the parent array, you should implement a subtype of TransferFunctions.IndexMapBorder or TransferFunctions.IndependentIndexMapBorder

TransferFunctions.IndependentIndexMapBorder — Type

IndependentIndexMapBorder{T} <: IndexMapBorder{T}

A border that maps the indices of the border independently for each dimension

TransferFunctions.IndexMapBorder — Type

IndexMapBorder{T} <: AbstractBorder{T}

A border that maps its values the values of the parent array with a certain index mapping

To implement the IndexBorderMap, you need to implement the method

Base.getindex — Method

mapindex(b::IndexMapBorder{T}, A::AbstractArray{T,N}, I::Vararg{Int,N}) where {T,N}

Map the BorderArray index to a parent array index with the given scheme.