Filtering

TransferFucntions provides some functions for discrete filtering. Both convolution and correlation have a dedicated function. This function accepts an image A, kernel K and an optional border and an output type T. There are both mutating and non mutating versions of the functions.

TransferFunctions.conv — Method

conv([T], A, K, [border])

Discrete convolution of A with K.

A is extended by the border. If T is given, the output will have the eltype T. If the border is not specified, no border is added, which for the FFT behaves the same as if :circular border was used.

See also conv!, corr, corr!

TransferFunctions.conv! — Function

conv!(out, A, K, [border])

Mutating version of conv.

See conv for details.

TransferFunctions.corr — Function

corr([T], A, K, [border])

Discrete correlation of A with K.

A is extended by the border. If T is given, the output will have the eltype T. If the border is not specified, no border is added, which for the FFT behaves the same as if :circular border was used.

See also corr!, conv, conv!

TransferFunctions.corr! — Function

corr!(out, A, K, [border])

Mutating version of corr.

See corr for details.