Airy Disc Model

The Airy disc describes the diffraction-limited point spread function (PSF) of a circular aperture, which is the fundamental resolution limit of a conventional microscope. It arises from scalar diffraction theory under the assumption of a perfectly aligned, aberration-free, incoherent imaging system.

It is parametrized by the wavelength ($\lambda$) of the emitted light, refractive index of the medium ($n$), and the numerical aperture ($\mathrm{NA}$).

TransferFunctions.AiryDisc — Type

AiryDisc{N, T<:Real} <: PSFModel{N}

Airy disc scalar diffraction point spread function model for a circular aperture.

Fields

λ::Length: emission wavelength

NA::Real: numerical aperture of the objective

n::Real: refractive index of the immersion medium (defaults to 4//3 which is the refractive index of water)

TransferFunctions.AiryDisc — Method

AiryDisc{N}(λ::Length, NA, n=4//3)

AiryDisc model point spread function with emission wavelength λ, numerical aperture NA and immersion medium refractive index n.

In the lateral plane it has the mathematical form of

\[h(r) \;=\; \left( \frac{2 J_1\!\left(\tfrac{\pi \, \mathrm{NA}}{\lambda} \, r\right)}{\tfrac{\pi \, \mathrm{NA}}{\lambda} \, r} \right)^{\!2},\]

where $J_1$ is the Bessel function of the first kind, $\lambda$ is the wavelength of the propagated light in the medium and $\mathrm{NA} = n \sin \theta$ is the numerical aperture.

TransferFunctions.intensity — Method

intensity(psf::AiryDisc, r::Length)

Compute the intensity of the AiryDisc PSF at a distance r from its center.

Intensity of the Airy disc is given by $h(r) = (2J₁(αr)/(αr))²$ where $α = 2πNA/λ$ and $J₁$ is the first order Bessel function of the first kind.

In the axial direction it is defined as proportionally to the lateral plane with the ratio

\[h(z) \;\propto\; \left( \frac{\sin\!\left(\tfrac{\pi \, \mathrm{NA}^2}{\lambda n} \, z\right)}{\tfrac{\pi \, \mathrm{NA}^2}{\lambda n} \, z} \right)^{\!2}.\]

TransferFunctions.axialintensity — Method

axialintensity(tf::AiryDisc{3}, z::Length)

Calculate the axial intensity of an AiryDisc PSF at the $z$-offset z.

Axial intensity of the Airy disc is given by $h(r) = \mathrm{sinc}(ϕ_z)²$ where $ϕ_z = NA²z/λn$.

TransferFunctions.intensity — Method

intensity(tf::AiryDisc{3}, r::Length, z::Length)

Compute the intensity of the AiryDisc PSF at a distance r from its center and $z$-offset z.

It is defined as the product of the lateral intensity and the axial intensity.

With the parameters

λ = 488u"nm"    # wavelength
NA = 1.4        # numerical aperture
n = 1.5         # refractive index of medium

you can define the AiryDisc PSF model for a 3D system as

airypsf_3d = AiryDisc(λ, NA, n)

AiryDisc{3, Float64}(488.0 nm, 1.4, 1.5)

or for a 2D system as

airypsf_2d = AiryDisc{2}(λ, NA)

AiryDisc{2, Float64}(λ=488.0 nm, NA=1.4)

Characteristics

Lateral FWHM (in the focal plane) is approximately equal to

\[\mathrm{FWHM}_\perp \;\approx\; 0.51 \,\frac{\lambda}{\mathrm{NA}}\]

which you can check by

FWHM_lateral = TF.FWHM(airypsf_3d)[1]

179.33894905055274 nm

which gives

FWHM_lateral * NA / λ

0.5144969849810939

The Axial FWHM (along the optical axis) is approximately given by

\[\mathrm{FWHM}_z \;\approx\; 0.885 \,\frac{\lambda n}{\mathrm{NA}^2}\]

and can be obtained with

FWHM_axial = TF.FWHM(airypsf_3d)[3]

330.8538944333461 nm

which gives

FWHM_axial * NA^2 / (λ * n)

0.8858929413789047

The energy for a given radius of the Airy disc has a closed form expression and can be computed using the method

TransferFunctions.encircled_energy — Method

encircled_energy(tf::AiryDisc, R::Length)

Compute the encircled energy of the Airy disc PSF tf for a circle of radius R.

An Airy disc has a closed from encircled energy of $E(R) = 1 - J₀²(αR) - J₁²(αR)$ where $α = 2πNA/λ$, $J₀$ and $J₁$ are the zeroth and first order Bessel functions of the first kind respectively.Born - Principles of Optics 8.5.2

See also energy_radius

TF.encircled_energy(airypsf_2d, 300u"nm")

0.8793766174490745

For a desired contained energy the correct radius can be found by bisection which and can be computed using the method

TransferFunctions.energy_radius — Method

energy_radius(tf::AiryDisc, ε::Real)

Calculate the energy radius of the Airy disc PSF tf for a given error term ε.

An Airy disc has an closed form encircled energy formula. The energy radius is computed by finding a root through bisection.

TF.energy_radius(airypsf_2d, 0.05)

682.4803099904603 nm