Born & Wolf Point Spread Function Model

The Born & Wolf model comes from the classical diffraction theory described in Principles of Optics and is thought of as the baseline diffraction-limited model used both in microscopy and astronomy.

TransferFunctions.BornWolf — Type

BornWolf{N,T<:Real} <: PSFModel{N}

Born & Wolf point spread function model.

Fields

λ::Length: emission wavelength

NA::Real: numerical aperture of the objective

n::Real: refractive index of the immersion medium (defaults to 4//3 which is the refractive index of water)

The Born & Wolf model is derived for a finite aperture ideal lens of a perfect system. It generalizes the mathematical formulation of the Airy pattern by allowing defocus and a different aperture shape. It assumes that the only aberration of the system is due to defocus. It assumes monochromatic light of a single wavelength, therefore chromatic effects are ignored. Modern microscope objectives are designed to provide optimal imaging conditions for sources located directly on the coverslip, in which case the Born & Wolf model is applicable (if the coverslip and immersion is used as designed). The model disregards spherical and higher order aberrations that are due to the source of illumination being shifted from the coverslip boundary. It is meant to be used for low to moderate $NA$ systems (ideally $NA <: 0.7$).

λ = 488u"nm"    # wavelength
NA = 1.4        # numerical aperture
n = 1.5         # refractive index of medium

bwpsf = BornWolf(λ, NA, n)

BornWolf{3, Float64}(488.0 nm, 1.4, 1.5, 1.3333333333333333)