Gaussian Point Spread Function Model

TransferFunctions.IsotropicGaussian — Type

IsotropicGaussian{T<:Real} <: PSFModel{3}

Airy disc scalar diffraction point spread function model for a circular aperture.

Fields

λ::Length: emission wavelength

NA::Real: numerical aperture of the objective

n::Real: refractive index of the immersion medium (defaults to 4//3 which is the refractive index of water)

TransferFunctions.C_AiryDisc_lateral — Constant

const C_AiryDisc_lateral

$c$ such that the FWHM of the Airy disc in the lateral direction is $c * λ/NA$.

TransferFunctions.C_AiryDisc_axial — Constant

const C_AiryDisc_axial

$c$ such that the FWHM of the Airy disc in the axial direction is $c * λn/NA^2$.

Characteristics

TransferFunctions.encircled_energy — Method

encircled_energy(tf::IsotropicGaussian, R::Length)

Compute the encircled energy of the Gaussian PSF tf for a circle of radius R.

An isotropic Gaussian has a closed from encircled energy of $E(R) = 1 - exp{-R² / 2σ²}$.

See also energy_radius

TransferFunctions.energy_radius — Method

energy_radius(tf::IsotropicGaussian, ε::Real)

Calculate the energy radius of the Gaussian PSF tf for a given error term ε.

An isotropic Gaussian has a closed form energy radius of $R(ε) = σ √(2 ln(1/ε))$.

See also encircled_energy