TransferFunctions.MeasuredPSF — Type

MeasuredPSF{T<:Real, N} <: PointSpreadFunction{N}

A measurement of a PSF fro the acquired image(s).

A measurement can be done for example using an image of subresolution microspheres.

Fields

data::AbstractArray{T,N}

Δxy::PixelSize{N}

center::Coordinate{N}

TransferFunctions.MeasuredOTF — Type

MeasuredOTF{T<:Number, N} <: OpticalTransferFunction{N}

A measurement of an OTF from the acquired image(s).

Fields

data::AbstractArray{T,N}

Δxy::PixelSize{N}

center::Coordinate{N}

Estimation

TransferFunctions.Estimation.bead — Function

bead([T=Float64], d, Δxy, [α=0u"nm^-1"],)
bead(d, α, Δxy)

Generate a model of a bead with diameter d and pixel-size Δxy.

Arguments

α = 1u"nm^-1: evanescent wave attenuation constant in the TIR-FM microscopy modulation. For a normal acquisition without total internal reflection 0u"nm^-1" is setting.

pixel_grid_length::Int = 10: length of each pixel in the grid

peak_intensity = 1.0: peak intensity value

subpixel_shift = (0.0, 0.0)

Examples

julia> Estimation.bead(100u"nm", 30.5u"nm");

julia> Estimation.bead(95u"nm", (30.5u"nm", 25u"nm"));

julia> Estimation.bead(Float32, 0.1u"μm", 30.5u"nm");

julia> Estimation.bead(0.1u"μm", 30.5u"nm"; α=0.01u"nm^-1")
5×5 OffsetArray(::Matrix{Float64}, -2:2, -2:2) with eltype Float64 with indices -2:2×-2:2:
 0.0        0.0       0.0705075  0.0       0.0
 0.0        0.606779  0.892914   0.606779  0.0
 0.0705075  0.892914  1.0        0.892914  0.0705075
 0.0        0.606779  0.892914   0.606779  0.0
 0.0        0.0       0.0705075  0.0       0.0

julia> Estimation.bead(0.1u"μm", 50.5u"nm"; subpixel_shift=(-0.3,-0.2))
3×3 OffsetArray(::Matrix{Float64}, -1:1, -1:1) with eltype Float64 with indices -1:1×-1:1:
 0.333333  0.737374  0.0808081
 0.59596   1.0       0.20202
 0.020202  0.141414  0.0

julia> Estimation.bead(0.1u"μm", 50.5u"nm"; peak_intensity=0.5)
3×3 OffsetArray(::Matrix{Float64}, -1:1, -1:1) with eltype Float64 with indices -1:1×-1:1:
 0.03  0.23  0.03
 0.23  0.5   0.23
 0.03  0.23  0.03