Unit Types

TransferFunctions.Frequency — Type

TransferFunctions.Frequency{T, U}

A supertype for quantities and levels of dimension Unitful.𝐋 ^ -1 with a value of type T and units U.

See also: Unitful.Quantity, Unitful.Level.

TransferFunctions.FrequencyUnits — Type

TransferFunctions.FrequencyUnits{U}

A supertype for units of dimension Unitful.𝐋 ^ -1. Equivalent to Unitful.Units{U, Unitful.𝐋 ^ -1}.

See also: Unitful.Units.

TransferFunctions.FrequencyFreeUnits — Type

TransferFunctions.FrequencyFreeUnits{U}

A supertype for Unitful.FreeUnits of dimension Unitful.𝐋 ^ -1. Equivalent to Unitful.FreeUnits{U, Unitful.𝐋 ^ -1}.

Array Types

TransferFunctions.CirculantTensor — Type

CirculantTensor{T,AA,N,KI,II} <: AbstractArray{T,N}

N-dimensional circulant tensor of an M-dimensional array A with kernel indices kern.

The dimensionality N is equal to 2M, where the first M dimensions have interior filtering coordinates ct.interior given by the kernel axes ct.kern and the axes of A and the tail M dimensions have the kernel coordinates ct.kern.

A correlation of array A with the kernel Kcan be obtained by outer tensor contraction over the tail M dimensions of the circulant(A,K) with K.

See also FilteringMatrix, BorderArray circulant

TransferFunctions.circulant — Function

circulant(A, K, border)

Construct a CirculantTensor of A with the kernel K such that tensor contraction with the front indices with K outputs the filtered array of A with K.

If border is specified, the array is padded with the strategy border so that the full extent of the array A is kept in the contraction output.

julia> circulant(reshape(1:25, (5,5)), (-1:1, -1:1));

julia> circulant(reshape(1:25, (5,5)), OAs.OffsetArray(ones(3,3), -1:1, -1:1));

julia> c = circulant(reshape(1:25, (5,5)), (-1:1, -1:1), :replicate);

julia> size(c)
(3, 3, 5, 5)

TransferFunctions.FilteringMatrix — Type

FilteringMatrix{T,M,AA,KI,II} <: AbstractMatrix{T}

A filtering matrix for the given M-dimensional array A and kernel indices kern. I.e. if K is an M-dimensional kernel with indices kern and F is the corresponding FilteringMatrix, then F * K[:] is the correlation of (or convolution if reflect(K)[:] where reflect(K) has the indices kern) of the K with A.

TransferFunctions.SampledArray — Type

SampledArray{T,ST,N,AA<:AbstractArray} <: AbstractArray{T,N}

An N-dimensional sampled array with data of type T sample single spacing of type ST.

The restriction of ST being a single type means that SampledArray is not able to represent arrays that have different dimensions in each direction.

See also SampledMatrix, SampledVector, SpatialArray

TransferFunctions.SampledMatrix — Type

SampledMatrix{T, ST} <: AbstractMatrix{T}

Two-dimensional array with elements of type T with a given sample spacing of type ST. Alias for SampledArray{T,ST,2}.

TransferFunctions.SampledVector — Type

SampledVector{T, ST} <: AbstractVector{T}

One-dimensional array with elements of type T with a given sample spacing of type ST. Alias for SampledArray{T,ST,1}.

TransferFunctions.SpatialArray — Type

SpatialArray{T,N} <: AbstractArray{T,N}

An N-dimensional array representing values sampled uniformly in space. Alias for SampledArray{T,<:Length,N}.

See also SpatialMatrix, SpatialVector, SampledArray

TransferFunctions.SpatialMatrix — Type

SpatialMatrix{T} <: AbstractMatrix{T}

Two-dimensional array with elements of type T with a given sampling distance in both directions. Alias for SpatialArray{T,2}.

TransferFunctions.SpatialVector — Type

SpatialVector{T} <: AbstractVector{T}

One-dimensional array with elements of type T with a given sampling distance. Alias for SpatialArray{T,1}.

Transfer Functions

TransferFunctions.TransferFunction — Type

TransferFunction

Super type for all transfer functions

A transfer function in microscopy is an object specifying the response in the image plane to the light coming from the object plane. There are several characteristics of a transfer functions which can influence the character of the response. Most commonly we are concerned with its spatial variance/invariance, i.e. whether the response is dependent of the objects position in the object plane, and linearity (which is almost always satisfied).

See also LinearTransferFunction, ImpulseResponseMapping

Linear Transfer Functions

TransferFunctions.LinearTransferFunction — Type

LinearTransferFunction <: TransferFunction

A supertype for all linear transfer functions.

A linear transfer function is that which ensures the system to have a linear response. This means that the system obeys the superposition principle, i.e. the response to a superposition of inputs is a superposition of the corresponding responses, and shift-invariance principle, i.e. the response to a signal is invariant to translation, which can be restated as "The single object in the object plane produces the same output in the image plane irrespective of its position within the object plane."

See also TransferFunctions

TransferFunctions.conv — Function

conv(ltf::LinearTransferFunction, img::SpatialArray{<:Real,2})

Transfer the image img using the linear transfer function ltf, i.e. convolve the image with the equivalent PSF.

conv(tf::PointSpreadFunction, img::SpatialArray{<:Real,2}, [args...]; <kwargs>)

Convolve the image img with the PSF tf. Additional arguments are passed to imfilter.

Keyword arguments

border=nothing If the border is a NamedTuple with the keys border of a type compatible with

BorderArray and apodization of type Apodization.ApodizationFunction, the border is applied to the image with the size of the FWHM of the PSF in the corresponding directions with the border and apodization used for edge tapering (taperedges)) and the full expanded array is returned.

TransferFunctions.deconv — Function

deconv(ltf::LinearTransferFunction, img::SpatialArray{<:Real,2})

Deconvolve the image img that was transferred using the linear transfer function ltf using the specified algorithm.

Optical Transfer Functions

TransferFunctions.OpticalTransferFunction — Type

OpticalTransferFunction <: LinearTransferFunction

Implementation

To create a new Optical transfer function (OTF) A <: OpticalTransferFunction, you must define the attenuation at a given frequency coordinate attenuation(otf::A, kx::Frequency, ky::Frequency).

TransferFunctions.OTFArray — Type

OTFArray{T<:Number} <: OpticalTransferFunction

Fields

data::AbstractMatrix{T}

Δxy::PixelSize{2}

center::Coordinate{2}

extend::ExtensionMethod

TransferFunctions.RadialOTF — Type

RadialOTF <: OpticalTransferFunction

If the pupil function of the system is symmetric, the OTF as well as the PSF are radially symmetric which can be used to optimize the calculations.

Implementation

attenuation(model::A, fr::Frequency)

cutoff(model::A, [a=0]) returning the largest frequency f such that attenuation(model, f) >= a if a > 0 and

attenuation(model, f) > 0 if a = 0.

TransferFunctions.CircularPupilOTF — Type

CircularPupilOTF(ρ₀::Real)

Ideal (aberration free) OTF of a diffraction limited imaging system with incoherent light with the cutoff-frequency ρ₀

TransferFunctions.otf — Function

otf(tf::OpticalTransferFunction, Δxy::PixelSize{2}, wh::Dims{2})

TransferFunctions.insupport — Function

insupport(otf::OpticalTransferFunction, fx::Frequency, fy::Frequency)

TransferFunctions.cutoff — Function

cutoff(::RadialOTF, a=0.0)

TransferFunctions.attenuation — Function

attenuation(otf::OpticalTransferFunction, ::Frequency, ::Frequency)

Point Spread Functions

TransferFunctions.PointSpreadFunction — Type

PointSpreadFunction <: LinearTransferFunction

A point spread function is a description of a transfer functions specifying the intensity transfer of a single point source in the object plane into a region of the image plane.

Implementation

To create a new Point spread function (PSF) A <: PointSpreadFunction, you must define the intensity at a given length coordinate intensity(psf::A, x::Length, y::Length).

TransferFunctions.MeasuredPSF — Type

MeasuredPSF <: PointSpreadFunction

TransferFunctions.PSFArray — Type

PSFArray{T<:Real} <: PointSpreadFunction

A measurement of a PSF from the acquired image(s).

Fields

data::AbstractMatrix{T}

Δ::PixelSize{2}

center::Coordinate{2}

extend::ExtensionMethod

TransferFunctions.ModelPSF — Type

ModelPSF <: PointSpreadFunction

TransferFunctions.RadialPSF — Type

RadialPSF <: ModelPSF

TransferFunctions.BornWolf — Type

BornWolf{T<:Real} <: RadialPSF

Born & Wolf model point spread function

Fields

+ `λ::Length{T}`: emission wavelength 
+ `n::T`: index of refraction of the immersion medium
+ `NA::T`: numerical aperture of the objective

TransferFunctions.GibsonLanni — Type

GibsonLanni{T<:Real} <: RadialPSF

Gibson & Lanni model point spread function

Fields

+ `λ::Length{T}`: emission wavelength
+ `NA::T`: numerical aperture of the objective
+ `n_i::T`: refractive index of the immersion medium
+ `n_s::T`: refractive index of the sample
+ `n_g::T`: refractive index of the coverslip
+ `t_i::Length{T}`: working distance of the objective
+ `t_g::Length{T}`: coverslip thickness

TransferFunctions.psf — Function

psf(tf::PointSpreadFunction, Δ::PixelSize{2}, wh::Dims{2})

Generate a PSF array size wh for the model tf with the pixel size Δ.

TransferFunctions.intensity — Function

intensity(psf::PointSpreadFunction, x::Length, y::Length)

The intensity of a PSF at a given location.

TransferFunctions.fit — Function

fit(::ModelPSF, ::SpatialArray)

TransferFunctions.FWHM — Function

FWHM(psf::PointSpreadFunction)

Find the FWHM of the PSF psf in the x and y directions.

julia> tf = TransferFunctions.BornWolf(λ=488u"nm", NA=1.4)
BornWolf{Float64}(488.0 nm, 1.4, 1.3333333333333333)

julia> TransferFunctions.FWHM(tf)
((-119.55929936703521 nm, 119.55929936703521 nm), (-119.55929936703521 nm, 119.55929936703521 nm))

Non-linear Transfer Functions

TransferFunctions.NonLinearTransferFunction — Type

NonLinearTransferFunction <: TransferFunction

A supertype for all non-linear transfer functions.

See also TransferFunctions, LinearTransferFunction

TransferFunctions.ImpulseResponseMapping — Type

ImpulseResponseMapping <: NonLinearTransferFunction

An impulse response mapping is a description of a transfer function that prescribes an impulse response to every point in the object plane. This may be a point spread function or some other function that prescribes the response to the object plane point in the image plane.

An transfer defined by an impulse response mapping is a linear system, but is not translation invariant hence does not belong to the class of linear transfer functions.

Estimation

TransferFunctions.Estimation.bead — Function

bead([T=Float64], d, Δ; <kwargs>)

Generate a model of a fluorescent microsphere (AKA calibration bead) with diameter d and pixel-size Δ.

Arguments

α=0u"nm^-1: evanescent wave attenuation constant in the TIR-FM microscopy modulation. An acquisition that is not TIR-FM is equivalent to setting α=0u"nm^-1".

pixel_grid=10: aliasing of the pixels is done by averaging values at a larger grid. pixel_grid sets the dimensions equivalent to each pixel in the grid

intesity=1.0: peak intensity value, i.e. the theoretical value at the center of the bead

position=(0.0, 0.0): set the subpixel position of the peak (center) of the bead within the generated array

Examples

julia> Estimation.bead(100u"nm", 30.5u"nm");

julia> Estimation.bead(95u"nm", (30.5u"nm", 25u"nm"));

julia> Estimation.bead(Float32, 0.1u"μm", 30.5u"nm");

julia> Estimation.bead(0.1u"μm", 30.5u"nm"; α=0.01u"nm^-1")
5×5 SampledArray{Float64, Quantity{Float64, 𝐋, Unitful.FreeUnits{(nm,), 𝐋, nothing}}, 2, OffsetArrays.OffsetMatrix{Float64, Matrix{Float64}}} with indices -2:2×-2:2:
 0.0        0.0       0.0705075  0.0       0.0
 0.0        0.606779  0.892914   0.606779  0.0
 0.0705075  0.892914  1.0        0.892914  0.0705075
 0.0        0.606779  0.892914   0.606779  0.0
 0.0        0.0       0.0705075  0.0       0.0

julia> Estimation.bead(0.1u"μm", 50.5u"nm"; position=(-0.3,-0.2))
3×3 SampledArray{Float64, Quantity{Float64, 𝐋, Unitful.FreeUnits{(nm,), 𝐋, nothing}}, 2, OffsetArrays.OffsetMatrix{Float64, Matrix{Float64}}} with indices -1:1×-1:1:
 0.333333  0.737374  0.0808081
 0.59596   1.0       0.20202
 0.020202  0.141414  0.0

julia> Estimation.bead(0.1u"μm", 50.5u"nm"; intensity=0.5)
3×3 SampledArray{Float64, Quantity{Float64, 𝐋, Unitful.FreeUnits{(nm,), 𝐋, nothing}}, 2, OffsetArrays.OffsetMatrix{Float64, Matrix{Float64}}} with indices -1:1×-1:1:
 0.03  0.23  0.03
 0.23  0.5   0.23
 0.03  0.23  0.03

TransferFunctions.Estimation.LeastSquares — Type

LeastSquares <: FromGroundTruth{PSFArray}

Fields

size::Size{2} – size of the PSF to estimate

Apodization

TransferFunctions.Apodization — Module

TransferFunctions.Apodization

Module defining apodization functions.

Apodization types Blackman, ExactBlackman, Connes, Cosine, Gaussian, Hamming, Welch, BlackmanNuttall, PowerCosine, Triangular, Nuttall, SineSum, BlackmanHarris, FlatTop and Hann

TransferFunctions.Apodization.ApodizationFunction — Type

ApodizationFunction{T}

Abstract type for apodization functions with return values of type T.

TransferFunctions.Apodization.Blackman — Type

Blackman{T,α} == SineSum{T,((1 - α) / 2, 1 / 2, α / 2)} <: ApodizationFunction{T}

Blackman apodization function with the coefficient α (Defaults to α = 0.16).

See SineSum See also ExactBlackman

TransferFunctions.Apodization.ExactBlackman — Type

ExactBlackman{T} == Blackman{T,683 // 4652}

See Blackman

TransferFunctions.Apodization.Connes — Type

Connes{T} <: ApodizationFunction{T}

TransferFunctions.Apodization.Cosine — Type

Cosine{T} == PowerCosine{T, 1} <: ApodizationFunction{T}

zero-phase function: $w₀(r) = cos(πr/2)$

instrument function: $I(k) = 4cos(2k)/(π(1 - 16k²))$

TransferFunctions.Apodization.Gaussian — Type

Gaussian{T} <: ApodizationFunction{T}

Gaussian apodization function with a given standard deviation σ.

zero-phase function: $w₀(r) = e^{-r²/2σ²}$

TransferFunctions.Apodization.Hamming — Type

Hamming{T} == SineSum{T,(25//46, 21//46)} <: ApodizationFunction{T}

See SineSum

TransferFunctions.Apodization.Welch — Type

Welch{T} <: ApodizationFunction{T}

zero-phase function: $w₀(r) = 1 - r²$

TransferFunctions.Apodization.BlackmanNuttall — Type

BlackmanNuttall{T} = SineSum{T,(0.3635819, 0.4891775, 0.1365995, 0.0106411)} <: ApodizationFunction{T}

See SineSum

TransferFunctions.Apodization.PowerCosine — Type

PowerCosine{T} <: ApodizationFunction{T}

zero-phase function: $w₀(r) = cos(πr/2)^α$

Instances: Cosine and Hann

TransferFunctions.Apodization.Triangular — Type

Triangular{T} <: ApodizationFunction{T}

Formulas:

zero-phase function: $w₀(r) = 1-|r|$

instrument function: $I(k) = sinc²(π k)$

TransferFunctions.Apodization.Nuttall — Type

Nuttall{T} == SineSum{T,(0.355768, 0.487396, 0.144232, 0.012604)} <: ApodizationFunction{T}

See SineSum

TransferFunctions.Apodization.SineSum — Type

SineSum{T,Cs} <: ApodizationFunction{T}

Sum of sines with coefficients Cs with output type of T

zero-phase function: $w₀(r) = ∑ᴺₖ₌₀ (-1)ᵏ Cs[k] cos(πk(r + 1))$

Instances: Hamming, Nuttall, BlackmanNuttall, BlackmanHarris, FlatTop, Blackman and ExactBlackman

TransferFunctions.Apodization.BlackmanHarris — Type

BlackmanHarris{T} == SineSum{T,(0.35875, 0.48829, 0.14128, 0.01168)} <: ApodizationFunction{T}

See SineSum

TransferFunctions.Apodization.FlatTop — Type

FlatTop{T} = SineSum{T,(0.21557895, 0.41663158, 0.277263158, 0.083578947, 0.006947365)} <: ApodizationFunction{T}

MATLAB variant of the flat-top filter

See SineSum

TransferFunctions.Apodization.Hann — Type

Hann{T} == PowerCosine{T,2} <: ApodizationFunction{T}

Formulas:

zero-phase function: $w₀(r) = cos²(πr/2)$

Internals

Functions

TransferFunctions.roundcenter — Function

roundcenter(r::RoundingMode, A)

Calculate the center index of A (::AbstractArray/::Indices/::Size/::AbstractUnitRange) rounded with the mode r::RoundingMode.

See also roundupcenter, rounddowncenter, exactcenter

TransferFunctions.roundupcenter — Function

roundupcenter(A)

Calculate center index of A rounded-up (i.e. fft center).

Examples

julia> TF.roundupcenter(ones(15,15))
CartesianIndex(8, 8)

julia> TF.roundupcenter((16,16))
CartesianIndex(9, 9)

TransferFunctions.rounddowncenter — Function

rounddowncenter(A)

Calculate center index of A rounded-down (i.e. ifft center).

Examples

julia> TF.rounddowncenter(ones(15,15))
CartesianIndex(8, 8)

julia> TF.rounddowncenter((16,16))
CartesianIndex(8, 8)

TransferFunctions.exactcenter — Function

 exactcenter(A)

Calculate the exact center coordinate of A.

Examples

julia> TF.exactcenter(ones(15,15))
(8.0, 8.0)

julia> TF.exactcenter((16,16))
(8.5, 8.5)

TransferFunctions.aroundorigin — Function

aroundorigin(s::Size, origin=(0,0,...))

Return the offset axes of an array of size s around the origin.

TransferFunctions.contained — Function

contained(A, loc::Coordinate{N})

Return true if the coordinate loc is contained in the axes of array A.

TransferFunctions.interior — Function

interior(inds::Indices{N}, kern::Indices{N})

Return 'valid' indices for convolution of an array with indices inds with a kernel having the indices kern.

TransferFunctions.fftfreqs — Function

fftfreqs(sz::Size{2}, Δ::PixelSize{2})
fftfreqs(sz::Size{2}, Δ::Length)

Generate a frequency grid for 2D FFTs with the pixel size Δ

julia> x_f, y_f = TF.fftfreqs((5,5), 50u"nm");

julia> x_f
5×5 Matrix{Quantity{Float64, 𝐋^-1, Unitful.FreeUnits{(nm^-1,), 𝐋^-1, nothing}}}:
  0.0 nm^-1     0.0 nm^-1     0.0 nm^-1     0.0 nm^-1     0.0 nm^-1
  0.004 nm^-1   0.004 nm^-1   0.004 nm^-1   0.004 nm^-1   0.004 nm^-1
  0.008 nm^-1   0.008 nm^-1   0.008 nm^-1   0.008 nm^-1   0.008 nm^-1
 -0.008 nm^-1  -0.008 nm^-1  -0.008 nm^-1  -0.008 nm^-1  -0.008 nm^-1
 -0.004 nm^-1  -0.004 nm^-1  -0.004 nm^-1  -0.004 nm^-1  -0.004 nm^-1

julia> y_f
5×5 Matrix{Quantity{Float64, 𝐋^-1, Unitful.FreeUnits{(nm^-1,), 𝐋^-1, nothing}}}:
 0.0 nm^-1  0.004 nm^-1  0.008 nm^-1  -0.008 nm^-1  -0.004 nm^-1
 0.0 nm^-1  0.004 nm^-1  0.008 nm^-1  -0.008 nm^-1  -0.004 nm^-1
 0.0 nm^-1  0.004 nm^-1  0.008 nm^-1  -0.008 nm^-1  -0.004 nm^-1
 0.0 nm^-1  0.004 nm^-1  0.008 nm^-1  -0.008 nm^-1  -0.004 nm^-1
 0.0 nm^-1  0.004 nm^-1  0.008 nm^-1  -0.008 nm^-1  -0.004 nm^-1

TransferFunctions.posgrid — Function

posgrid(sz::Size{2}, Δ::PixelSize{2}; center=roundupcenter(sz))

Generate a position grid for 2D sampled images with the pixel size Δ

julia> x_p, y_p = TF.posgrid((5, 5), 50u"nm");

julia> x_p
5×5 Matrix{Quantity{Float64, 𝐋, Unitful.FreeUnits{(nm,), 𝐋, nothing}}}:
 -100.0 nm  -100.0 nm  -100.0 nm  -100.0 nm  -100.0 nm
  -50.0 nm   -50.0 nm   -50.0 nm   -50.0 nm   -50.0 nm
    0.0 nm     0.0 nm     0.0 nm     0.0 nm     0.0 nm
   50.0 nm    50.0 nm    50.0 nm    50.0 nm    50.0 nm
  100.0 nm   100.0 nm   100.0 nm   100.0 nm   100.0 nm

julia> y_p
5×5 Matrix{Quantity{Float64, 𝐋, Unitful.FreeUnits{(nm,), 𝐋, nothing}}}:
 -100.0 nm  -50.0 nm  0.0 nm  50.0 nm  100.0 nm
 -100.0 nm  -50.0 nm  0.0 nm  50.0 nm  100.0 nm
 -100.0 nm  -50.0 nm  0.0 nm  50.0 nm  100.0 nm
 -100.0 nm  -50.0 nm  0.0 nm  50.0 nm  100.0 nm
 -100.0 nm  -50.0 nm  0.0 nm  50.0 nm  100.0 nm

TransferFunctions.fillsize — Function

fillsize(Δ::Length, N::Integer) => PixelSize{N}

Construct a PixelSize with the same sampling Δ in every direction.

Type Aliases

TransferFunctions.Size — Type

Size{N}

N-dimensional size of an array. Alias for NTuple{N,Int}

TransferFunctions.Coordinate — Type

Coordinate{N,T}

N-dimensional coordinate. Alias for NTuple{N,T<:Real}

TransferFunctions.PixelSize — Type

PixelSize{N}

N-dimensional pixel size. Alias for NTuple{N,Length}

To Sort

TransferFunctions.IndependentIndexMapBorder — Type

IndependentIndexMapBorder{T} <: IndexMapBorder{T}

A border that maps the indices of the border independently for each dimension

TransferFunctions.filtering_matrix — Function

filtering_matrix(A, K, [border])

Construct a FilteringMatrix of A with the kernel K such that the F'*K[:] produces a filtered vector of A.

If border is specified, the array is padded with the strategy border so that the full extent of the array A is kept in the filtering output.

julia> filtering_matrix(reshape(1:25, (5,5)), (-1:1, -1:1));

julia> filtering_matrix(reshape(1:25, (5,5)), OAs.OffsetArray(ones(3,3), -1:1, -1:1));

julia> filtering_matrix(reshape(1:9, (3,3)), (-1:1, -1:1), :symmetric)
9×9 filtering_matrix(border_array(reshape(::UnitRange{Int64}, 3, 3), :Symmetric), (3, 3)) with eltype Int64:
 5  4  5  2  1  2  5  4  5
 4  5  6  1  2  3  4  5  6
 5  6  5  2  3  2  5  6  5
 2  1  2  5  4  5  8  7  8
 1  2  3  4  5  6  7  8  9
 2  3  2  5  6  5  8  9  8
 5  4  5  8  7  8  5  4  5
 4  5  6  7  8  9  4  5  6
 5  6  5  8  9  8  5  6  5

TransferFunctions.AbstractBorder — Type

AbstractBorder{T}

Super type for borders of arrays with the element type T

Implementation

A new border type needs to implement

Base.getindex(b::AbstractBorder{T}, A::AbstractArray{T,N}, I::Vararg{Int,N}) where {T,N}

where A is the parent array of the BorderArray.

TransferFunctions.IndexMapBorder — Type

IndexMapBorder{T} <: AbstractBorder{T}

A border that maps its values the values of the parent array with a certain index mapping

TransferFunctions.Reflect — Type

Reflect{T} <: IndexMapBorder{T}

A border that reflects the values of the border around the edges of the parent array

julia> TransferFunctions.BorderArray(reshape(1:16, (4,4)), TransferFunctions.Reflect, 2)
8×8 border_array(reshape(::UnitRange{Int64}, 4, 4), :Reflect) with eltype Int64 with indices -1:6×-1:6:
 6  2  2  6  10  14  14  10
 5  1  1  5   9  13  13   9
 5  1  1  5   9  13  13   9
 6  2  2  6  10  14  14  10
 7  3  3  7  11  15  15  11
 8  4  4  8  12  16  16  12
 8  4  4  8  12  16  16  12
 7  3  3  7  11  15  15  11

TransferFunctions.kern_padding — Function

kern_padding(kern)

Determine the padding necessary to keep the input array fully contained in the interior of the output when filtered with kern.

TransferFunctions.TaperedArray — Type

TaperedArray{T,N,AA,IA,A} <: AbstractArray{T,N}

An N-dimensional array with data of type T that has edges of the width .edges tapered with an ApodizationFunction.

It is a lazy light wrapper that only computes the tapering when indexed.

It should be constructed using taperedges.

TransferFunctions.Fill — Type

Fill{T} <: AbstractBorder{T}

A border that fills the values with value::T

TransferFunctions.border_array — Function

border_array(A, border, padding)

Construct a BorderArray of A with the border border and padding padding.

julia> border_array(reshape(1:9, (3,3)), :circular, 2)
7×7 border_array(reshape(::UnitRange{Int64}, 3, 3), :Circular) with eltype Int64 with indices -1:5×-1:5:
 5  8  2  5  8  2  5
 6  9  3  6  9  3  6
 4  7  1  4  7  1  4
 5  8  2  5  8  2  5
 6  9  3  6  9  3  6
 4  7  1  4  7  1  4
 5  8  2  5  8  2  5

TransferFunctions.taperedges — Function

taperedges([apo=Cosine()], A, w, [border])
taperedges([apo=Cosine()], A, (w1,w2...,wn), [border])
taperedges([apo=Cosine()], A, ((w1a,w1b),...,(wna,wnb), [border])

Taper edges of width w of array A using apo::ApodizationFunction.

See also BorderArray, TaperedArray, ApodizationFunction

TransferFunctions.ind2sub — Function

ind2sub(inds::DimsInteger, ind::Integer)

Convert a linear index ind to a subscript index for the array size inds.

TransferFunctions.Circular — Type

Circular{T} <: IndexMapBorder{T}

A border that wraps the values around the edges of the parent array

julia> TransferFunctions.BorderArray(reshape(1:16, (4,4)), TransferFunctions.Circular, 2)
8×8 border_array(reshape(::UnitRange{Int64}, 4, 4), :Circular) with eltype Int64 with indices -1:6×-1:6:
 11  15  3  7  11  15  3  7
 12  16  4  8  12  16  4  8
  9  13  1  5   9  13  1  5
 10  14  2  6  10  14  2  6
 11  15  3  7  11  15  3  7
 12  16  4  8  12  16  4  8
  9  13  1  5   9  13  1  5
 10  14  2  6  10  14  2  6

TransferFunctions.InvalidBorderExtent — Type

InvalidBorderExtent <: Exception

Is thrown when trying to construct a BorderArray with padding larger than the maximum padding that would produce valid values.

TransferFunctions.Replicate — Type

Replicate{T}  <: IndexMapBorder{T}

A border that replicates the closest edge value of the parent array

julia> TransferFunctions.BorderArray(reshape(1:16, (4,4)), TransferFunctions.Replicate, 2)
8×8 border_array(reshape(::UnitRange{Int64}, 4, 4), :Replicate) with eltype Int64 with indices -1:6×-1:6:
 1  1  1  5   9  13  13  13
 1  1  1  5   9  13  13  13
 1  1  1  5   9  13  13  13
 2  2  2  6  10  14  14  14
 3  3  3  7  11  15  15  15
 4  4  4  8  12  16  16  16
 4  4  4  8  12  16  16  16
 4  4  4  8  12  16  16  16

TransferFunctions.Symmetric — Type

Symmetric{T} <: IndexMapBorder{T}

A border that symmetrically continues the values or the parent array around the edges, i.e. mirrors around the last valid parent index.

julia> TransferFunctions.BorderArray(reshape(1:16, (4,4)), TransferFunctions.Symmetric, 2)
8×8 border_array(reshape(::UnitRange{Int64}, 4, 4), :Symmetric) with eltype Int64 with indices -1:6×-1:6:
 11  7  3  7  11  15  11  7
 10  6  2  6  10  14  10  6
  9  5  1  5   9  13   9  5
 10  6  2  6  10  14  10  6
 11  7  3  7  11  15  11  7
 12  8  4  8  12  16  12  8
 11  7  3  7  11  15  11  7
 10  6  2  6  10  14  10  6

TransferFunctions.BorderArray — Type

BorderArray{T,N,AA,AB} <: AbstractArray{T,N}

A border array with a border of type AB with the parent array of type AA.

It is a light wrapper around the parent array that defines Base.getindex to return values with its supplied border scheme.

julia> TransferFunctions.BorderArray(reshape(1:36, (6,6)), TransferFunctions.Fill, 2)
10×10 border_array(reshape(::UnitRange{Int64}, 6, 6), fill(0)) with eltype Int64 with indices -1:8×-1:8:
 0  0  0   0   0   0   0   0  0  0
 0  0  0   0   0   0   0   0  0  0
 0  0  1   7  13  19  25  31  0  0
 0  0  2   8  14  20  26  32  0  0
 0  0  3   9  15  21  27  33  0  0
 0  0  4  10  16  22  28  34  0  0
 0  0  5  11  17  23  29  35  0  0
 0  0  6  12  18  24  30  36  0  0
 0  0  0   0   0   0   0   0  0  0
 0  0  0   0   0   0   0   0  0  0