User Interface

Any transfer function is a subtype of the abstract type TransferFunction

TransferFunctions.TransferFunction — Type

TransferFunction

Super type for all transfer functions

A transfer function in microscopy is an object specifying the response in the image plane to the light coming from the object plane. There are several characteristics of a transfer functions which can influence the character of the response. Most commonly we are concerned with its spatial variance/invariance, i.e. whether the response is dependent of the objects position in the object plane, and linearity (which is almost always satisfied).

See also LinearTransferFunction, ImpulseResponseMapping

TransferFunctions.OpticalTransferFunction — Type

OpticalTransferFunction <: LinearTransferFunction

Implementation

To create a new Optical transfer function (OTF) A <: OpticalTransferFunction, you must define the attenuation at a given frequency coordinate attenuation(otf::A, kx::Frequency, ky::Frequency).

TransferFunctions.PointSpreadFunction — Type

PointSpreadFunction <: LinearTransferFunction

A point spread function is a description of a transfer functions specifying the intensity transfer of a single point source in the object plane into a region of the image plane.

Implementation

To create a new Point spread function (PSF) A <: PointSpreadFunction, you must define the intensity at a given length coordinate intensity(psf::A, x::Length, y::Length).

You can get a transfer function of your optical setup by supplying parameters of the apparatus to a model transfer function that is developed from the underlining physics of a microscope, then you will use a subtype of Otherwise you can estimate the transfer function, most commonly by the means of an acquisition where the imaged sample is known such as sub-diffraction sized microspheres of known sizes.