Estimation

Estimation methods for transfer functions are implemented in the Estimation module of the TransferFucntions package.

Non-Blind Methods

Non-blind are the methods for estimation that require a pair or a set of pairs of ground truth images along with their corresponding acquired raw images (ones that are noisy ground truth images blurred by the transfer function).

Ground Truth

The ground truth must often also be estimated using a model of the acquired scene. One option for the pair could be an acquisition of sub-diffraction microspheres or polymer fluorescent beads along with a model consisting of the beads known dimensions and estimated positions in the scene.

TransferFunctions.Estimation.bead — Function

bead([T=Float64], d, Δ; <kwargs>)

Generate a model of a fluorescent microsphere (AKA calibration bead) with diameter d and pixel-size Δ.

Arguments

α=0u"nm^-1: evanescent wave attenuation constant in the TIR-FM microscopy modulation. An acquisition that is not TIR-FM is equivalent to setting α=0u"nm^-1".

pixel_grid=10: aliasing of the pixels is done by averaging values at a larger grid. pixel_grid sets the dimensions equivalent to each pixel in the grid

intesity=1.0: peak intensity value, i.e. the theoretical value at the center of the bead

position=(0.0, 0.0): set the subpixel position of the peak (center) of the bead within the generated array

Examples

julia> Estimation.bead(100u"nm", 30.5u"nm");

julia> Estimation.bead(95u"nm", (30.5u"nm", 25u"nm"));

julia> Estimation.bead(Float32, 0.1u"μm", 30.5u"nm");

julia> Estimation.bead(0.1u"μm", 30.5u"nm"; α=0.01u"nm^-1")
5×5 OffsetArray(::Matrix{Float64}, -2:2, -2:2) with eltype Float64 with indices -2:2×-2:2:
 0.0        0.0       0.0705075  0.0       0.0
 0.0        0.606779  0.892914   0.606779  0.0
 0.0705075  0.892914  1.0        0.892914  0.0705075
 0.0        0.606779  0.892914   0.606779  0.0
 0.0        0.0       0.0705075  0.0       0.0

julia> Estimation.bead(0.1u"μm", 50.5u"nm"; position=(-0.3,-0.2))
3×3 OffsetArray(::Matrix{Float64}, -1:1, -1:1) with eltype Float64 with indices -1:1×-1:1:
 0.333333  0.737374  0.0808081
 0.59596   1.0       0.20202
 0.020202  0.141414  0.0

julia> Estimation.bead(0.1u"μm", 50.5u"nm"; intensity=0.5)
3×3 OffsetArray(::Matrix{Float64}, -1:1, -1:1) with eltype Float64 with indices -1:1×-1:1:
 0.03  0.23  0.03
 0.23  0.5   0.23
 0.03  0.23  0.03