Transfer Functions

Linear Transfer Functions

PSF and OTF

Optical Transfer Functions (OTFs)

Circular Pupil Model

TransferFunctions.CircularPupilOTF — Type

CircularPupilOTF(ρ₀::Real)

Ideal (aberration free) OTF of a diffraction limited imaging system with incoherent light with the cutoff-frequency ρ₀

The OTF is derived from the diffraction caused by the exit pupil of the system and disregards the effect of the entrance pupil... thus assumes no reshaping of the wavefronts in the optical system. The exit pupil, being located in the optical system just before the light reaches the image plane, has a greater effect on the optical system OTF.

The OTF can be written in as a function of the cutoff frequency $ρ₀$ [1]

\[ ℋ(ρ) = \begin{cases} (2/π) \left\{ \arccos(ρ/2ρ₀) - (ρ/2ρ₀)\sqrt{1 - (ρ/2ρ₀)²} \right\} & \text{ for } ρ ≤ 2ρ₀ \\ 0 & \text{ otherwise}. \end{cases} \]

The cutoff frequency can be written in terms of the wavelength $λ$, distance between entrance pupil and the image plane $f₂$ and the circular pupil radius $w$ as

\[ ρ_0 = w/(λ f₂).\]

Point Spread Functions (PSFs)

Born & Wolf

TransferFunctions.BornWolf — Type

BornWolf{T<:Real} <: RadialPSF

Born & Wolf model point spread function

Fields

+ `λ::Length{T}`: emission wavelength 
+ `n::T`: index of refraction of the immersion medium
+ `NA::T`: numerical aperture of the objective

The Born & Wolf model is a scalar diffraction model derived for perfect systems. It assumes that the only aberration of the system is due to defocus. Modern microscope objectives are designed to provide optimal imaging conditions for sources located directly on the coverslip, in which case the Born & Wolf model is applicable (if the coverslip and immersion is used as designed). The model disregards spherical and higher order aberrations that are due to the source of illumination being shifted from the coverslip boundary.

TransferFunctions.BornWolf — Method

BornWolf(λ::Length, NA, n_i=4//3)

BornWolf model point spread function with emission wavelength λ, numerical aperture NA and immersion medium refractive index n_i.

Gibson & Lanni

TransferFunctions.GibsonLanni — Type

GibsonLanni{T<:Real} <: RadialPSF

Gibson & Lanni model point spread function

Fields

+ `λ::Length{T}`: emission wavelength
+ `NA::T`: numerical aperture of the objective
+ `n_i::T`: refractive index of the immersion medium
+ `n_s::T`: refractive index of the sample
+ `n_g::T`: refractive index of the coverslip
+ `t_i::Length{T}`: working distance of the objective
+ `t_g::Length{T}`: coverslip thickness

The Gibson & Lanni model is assumes that, disregarding defocus, all observed aberrations are generated by factors external to the objective (i.e. originating in the sample, coverslip and immersion medium combination). These aberrations can be characterized by the optical path difference between a ray in a perfect system (see BornWolf) and a ray under experimental conditions.

TransferFunctions.GibsonLanni — Method

GibsonLanni(λ::Length, NA, n_i=1.5, n_s=4//3, n_g=1.5, t_i, t_g)

GibsonLanni model point spread function with emission wavelength λ, numerical aperture NA, immersion medium refractive index n_i, sample refractive index n_s, coverslip refractive index n_g, working distance t_i and coverslip thickness t_g.