OTF Models

TransferFunctions.ModelOTF — Type

ModelOTF{N} <: OpticalTransferFunction{N}

An N-dimensional optical transfer function based on a physical model (in contrast to a measurement).

See also CircularPupilOTF

Implementation

To create a new OTF model A <: ModelOTF{2}, you must define the attenuation (transfer coefficient) at a given frequency:

attenuation(model::A, kx::Frequency, ky::Frequency) for a non-symmetric OTF, or

attenuation(model::A, kᵣ::Frequency) for a radially symmetric OTF and use @traitimpl RadiallySymmetric{A} to mark the symmetry trait implementation.

Optionally implement the methods:

preferred_type(::Type{<:ModelOTF})::T which returns the natural type T that is returned from calling the model. It should be either T<:Complex or T<:Real. Defaults to Float64.

For a symmetric OTF, you should consider defining:

cutoff(model::A, [a=0]) returning the largest frequency f such that attenuation(model, f) >= a if a > 0 and attenuation(model, f) > 0 if a = 0.

TransferFunctions.CircularPupilOTF — Type

CircularPupilOTF(ρ₀::Real)

Ideal (aberration free) OTF of a diffraction limited imaging system with incoherent light with the cutoff-frequency ρ₀

The OTF is derived from the diffraction caused by the exit pupil of the system and disregards the effect of the entrance pupil... thus assumes no reshaping of the wavefronts in the optical system. The exit pupil, being located in the optical system just before the light reaches the image plane, has a greater effect on the optical system OTF.

Extended help

The OTF can be written in as a function of the cutoff frequency $ρ₀$ [1]

\[ ℋ(ρ) = \begin{cases} (2/π) \left\{ \arccos(ρ/2ρ₀) - (ρ/2ρ₀)\sqrt{1 - (ρ/2ρ₀)²} \right\} & \text{ for } ρ ≤ 2ρ₀ \\ 0 & \text{ otherwise}. \end{cases} \]

The cutoff frequency can be written in terms of the wavelength $λ$, distance between entrance pupil and the image plane $f₂$ and the circular pupil radius $w$ as

\[ ρ_0 = w/(λ f₂).\]

References

[1] Goodman, J.W., 2005. Frequency Analysis of Optical Imaging Systems, in: Introduction to Fourier Optics. Roberts & Co, Englewood, Colo, pp. 127–172.

PSF Models

TransferFunctions.BornWolf — Type

Born & Wolf model of the transfer function for a circular aperture.

Parameters

λ::Unitful.Length: (emission) wavelength

nᵢ::Number: index of refraction of the immersion medium

NA::Number: numerical aperture

Extended help

The Born & Wolf model is a scalar diffraction model derived for perfect systems. It assumes that the only aberration of the system is due to defocus. Modern microscope objectives are designed to provide optimal imaging conditions for sources located directly on the coverslip, in which case the Born & Wolf model is applicable (if the coverslip and immersion is used as designed). The model disregards spherical and higher order aberrations that are due to the source of illumination being shifted from the coverslip boundary.

TransferFunctions.GibsonLanni — Type

Gibson & Lanni model of the transfer function for a circular aperture.

Parameters

Extended help

The Gibson & Lanni model is assumes that, disregarding defocus, all observed aberrations are generated by factors external to the objective (i.e. originating in the sample, coverslip and immersion medium combination). These aberrations can be characterized by the optical path difference between a ray in a perfect system (see BornWolf) and a ray under experimental conditions.