<!– FIX: These documentation pages should be textual and structured by concepts. The API reference is then for the API structured documentation. Types, functions, model types etc. These could be also organized by some key but they should mostly obey the their programming structure <30-03-25> –>

There are two main categories of transfer function types that are provided by this package

Subtypes of TransferFunction which define a general transfer function model/measurement of a microscope apparatus and

Subtypes of SampledTransferFunction that may are created by providing a sampling scheme (most commonly a pixel size) to the TransferFunction. These are used when you are working with specific acquisition with known sampling rates/schemes.

Transfer Functions

TransferFunctions.TransferFunctions — Module

Package for models, estimation, sampling and deconvolution with microscopy transfer functions.

You can describe a transfer of a microscope with either the optical transfer function (OTF) the point spread function (PSF) or the generalized pupil function. These descriptions are mostly interchangeable/convertible between each other.

Note

It may the case that a model can be in a closed form expression for some of these and must be approximated for the others, which may be a reason for the choice between them for a specific model.

TransferFunctions.OpticalTransferFunction — Type

OpticalTransferFunction{N} <: TransferFunction{N}

Super-type for any N-dimensional optical transfer function

See also PointSpreadFunction, TransferFunction

TransferFunctions.PointSpreadFunction — Type

PointSpreadFunction{N} <: TransferFunction{N}

Super-type for any N-dimensional point spread function

See also OpticalTransferFunction, TransferFunction

You can get a transfer function of your optical setup by supplying parameters of the apparatus to a model transfer function that is developed from the underlining physics of a microscope, then you will use a subtype of

TransferFunctions.ModelTransferFunction — Type

Union type for a transfer function that is based on a physical model of an optical system. Contrary to a MeasuredTransferFunction, a ModelTransferFunction must be quantifiable at any point (in either the spatial or the frequency domain).

Otherwise you can estimate the transfer function, most commonly by the means of an acquisition where the imaged sample is known such as sub-diffraction sized microspheres of known sizes. This approach will lead to a MeasuredTransferFunction

TransferFunctions.MeasuredTransferFunction — Type

A Union type for the measurement of a transfer function of an optical system. It can be either a PSF measurement MeasuredPSF or a MeasuredOTF.

Sampled Transfer Functions

<!– TODO: <18-11-24> –>