There are two main categories of transfer function types that are provided by this package

Subtypes of TransferFunction and

Subtypes of SampledTransferFunction.

The former facilitate the models or measured transfer functions that describe the transfer of a microscope apparatus. The latter can be generated from the former by supplying the parameters of the sampling process, most commonly the pixel size. A SampledTransferFunction is most commonly what you would want for processing data from your acquisition.

Transfer Functions

You can describe a transfer of a microscope with either the optical transfer function (OTF) the point spread function (PSF) or the generalized pupil function. These descriptions are mostly interchangeable/convertible between each other. However note, that it may the case that a model can be in a closed form expression for some of these and must be approximated for the others, which may be a reason for a choice between them.

You can get a transfer function of your optical setup by supplying parameters to a model transfer function that is developed from the underlining physics of a microscope, then you will use a subtype of

TransferFunctions.ModelTransferFunction — Type

Union type for transfer function that is based on a physical model of an optical system. Contrary to a MeasuredTransferFunction, a ModelTransferFunction must be quantifiable at any point (in either spatial or frequency domain).

or you can estimate the transfer function, most commonly by the means of an acquisition where the imaged sample is known such as sub-diffraction sized microspheres of known

TransferFunctions.MeasuredTransferFunction — Type

An abstract type for the measurement of the transfer function of an optical system. It can be either a PSF measurement MeasuredPSF or a MeasuredOTF.

Sampled Transfer Functions