Abstract Array Subtypes

General Storage Arrays

TransferFunctions.Flattened — Type

Flattened{P,OM,IM,AX,IA,T,N} <: AbstractArray{T,N}

An N-dimensional AbstractArray stored as an M-dimensional 'outer' parent array ::P of K-dimensional T-valued 'inner' arrays, where N=M+K.

It should be constructed by flatten.

The constructor ensures the inner arrays have the same axes. In a sense, this can be thought as the inverse of Slices from Base, which allows you to view a single array as multiple arrays. Flattened lets you see multiple arrays that are nested in one container array as a single flat array.

parent(f::Flattened) will return the nested parent array.

See also Slices

Fields

outermap::OM and innermap::IM are M and K integer long tuples respectively of the dimensions that the outer and inner arrays represent. They are ensured to be unique and to comprise the total of 1:N dimensions. axes::AX contains the axes of the flattened array. parent::P holds the nested array.

Examples

julia> flatten(eachslice(reshape(1:12,3,4); dims=1)) == reshape(1:12,3,4)
true

A container array that is used in the implementation of other array types that holds some dimensions in an inner array and these

Filtering Arrays

TransferFunctions.CirculantTensor — Type

CirculantTensor{P,M,AA,T,N} <: AbstractArray{T,N}

N-dimensional circulant tensor of an M-dimensional array A with kernel indices kern.

The dimensionality N is equal to 2M, where the first M dimensions have interior filtering coordinates ct.interior given by the kernel axes ct.kern and the axes of A and the tail M dimensions have the kernel coordinates ct.kern.

A correlation of array A with the kernel Kcan be obtained by outer tensor contraction over the tail M dimensions of the circulant(A,K) with K.

parent(ct::CirculantTensor) will return the Flattened array of views inside of A.

See also FilteringMatrix, imfilter, circulant

TransferFunctions.FilteringMatrix — Type

FilteringMatrix{P,M,CT,T} <: AbstractMatrix{T}

A filtering matrix for the given M-dimensional array A and kernel indices kern. I.e. if K is an M-dimensional kernel with indices kern and F is the corresponding FilteringMatrix, then F * K[:] is the correlation of (or convolution if reflect(K)[:] where reflect(K) has the indices kern) of the K with A.

It is implemented as a stacked view into the CirculantTensor of A for the kernel indices kern.

Data Arrays

For representing data that is sampled at some constant rate, possibly different in every dimension, there are SampledArrays that store the data coupled with its sampling.

TransferFunctions.SampledArray — Type

SampledArray{T,ST,N,AA<:AbstractArray} <: AbstractArray{T,N}

An N-dimensional sampled array with data of type T sample single spacing of type ST.

The restriction of ST being a single type means that SampledArray is not able to represent arrays that have different dimensions in each direction.

See also SampledMatrix, SampledVector, SpatialArray

TransferFunctions.SampledMatrix — Type

SampledMatrix{T, ST} <: AbstractMatrix{T}

Two-dimensional array with elements of type T with a given sample spacing of type ST. Alias for SampledArray{T,ST,2}.

TransferFunctions.SampledVector — Type

SampledVector{T, ST} <: AbstractVector{T}

One-dimensional array with elements of type T with a given sample spacing of type ST. Alias for SampledArray{T,ST,1}.

Since sampling at spatially defined lattice points is so common in the context of microscopy, constants for SampledArrays with spatial sampling are provided as aliases.

TransferFunctions.SpatialArray — Type

SpatialArray{T,N} <: AbstractArray{T,N}

An N-dimensional array representing values sampled uniformly in space. Alias for SampledArray{T,<:Length,N}.

See also SpatialMatrix, SpatialVector, SampledArray

TransferFunctions.SpatialMatrix — Type

SpatialMatrix{T} <: AbstractMatrix{T}

Two-dimensional array with elements of type T with a given sampling distance in both directions. Alias for SpatialArray{T,2}.

TransferFunctions.SpatialVector — Type

SpatialVector{T} <: AbstractVector{T}

One-dimensional array with elements of type T with a given sampling distance. Alias for SpatialArray{T,1}.