Unit Types

TransferFunctions.Frequency — Type

TransferFunctions.Frequency{T, U}

A supertype for quantities and levels of dimension Unitful.𝐋 ^ -1 with a value of type T and units U.

See also: Unitful.Quantity, Unitful.Level.

TransferFunctions.FrequencyUnits — Type

TransferFunctions.FrequencyUnits{U}

A supertype for units of dimension Unitful.𝐋 ^ -1. Equivalent to Unitful.Units{U, Unitful.𝐋 ^ -1}.

See also: Unitful.Units.

TransferFunctions.FrequencyFreeUnits — Type

TransferFunctions.FrequencyFreeUnits{U}

A supertype for Unitful.FreeUnits of dimension Unitful.𝐋 ^ -1. Equivalent to Unitful.FreeUnits{U, Unitful.𝐋 ^ -1}.

Array Types

TransferFunctions.OuterInnerArray — Type

OuterInnerArray{T,N,M,K,IA,OA} <: AbstractArray{T,N}

N-dimensional array that consists of an M dimensional 'outer' array of type OA<:AbstractArray{IA,M} of K dimensional T valued 'inner' arrays of type IA{T,K}. The constructor ensures that M+K==N and that the inner arrays have the same axes.

See also Slices

Fields

outer::OA

isinnerdim::SVector{N,Bool}

size::Size{N}

axes::Indices{N}

TransferFunctions.CirculantTensor — Type

CirculantTensor{T,N,M,AA} <: AbstractArray{T,N}

N-dimensional circulant tensor of an M dimensional array A or type AA<:AbstractArray{T,M}. The dimensionality N is equal to 2M, where the first M dimensions have interior filtering coordinates ct.interior given by the kernel axes ct.kern and the axes of A and the tail M dimensions have the kernel coordinates ct.kern.

A correlation filtering result of A with a kernel array Kcan be obtained by outer tensor contraction over the tail M dimensions of the CirculantTensor(A,K).

See also FilteringMatrix, imfilter

TransferFunctions.FilteringMatrix — Type

FilteringMatrix{T,M,P} <: AbstractMatrix{T}

A matrix that for a given kernel K, array A of size size(A)==size(K) equal to M and a padding scheme P gives an array F such that filtering (correlation or convolution if reflect(K) is used as the kernel) output of the kernel K and an array as A can be computed as F * A[:]. It is a view into the CirculantTensor for the kernel array K.

TransferFunctions.SampledArray — Type

SampledArray{T,ST,N,AA<:AbstractArray} <: AbstractArray{T,N}

An N-dimensional sampled array with data of type T sample spacing of type ST.

See also SampledMatrix, SampledVector, SpatialArray

TransferFunctions.SampledMatrix — Type

SampledMatrix{T, ST} <: AbstractMatrix{T}

Two-dimensional array with elements of type T with a given sample spacing of type ST. Alias for SampledArray{T,ST,2}.

TransferFunctions.SampledVector — Type

SampledVector{T, ST} <: AbstractVector{T}

One-dimensional array with elements of type T with a given sample spacing of type ST. Alias for SampledArray{T,ST,1}.

TransferFunctions.SpatialArray — Type

SpatialArray{T,N} <: AbstractArray{T,N}

An N-dimensional array representing values sampled uniformly in space. Alias for SampledArray{T,<:Length,N}.

See also SpatialMatrix, SpatialVector, SampledArray

TransferFunctions.SpatialMatrix — Type

SpatialMatrix{T} <: AbstractMatrix{T}

Two-dimensional array with elements of type T with a given sampling distance in both directions. Alias for SpatialArray{T,2}.

TransferFunctions.SpatialVector — Type

SpatialVector{T} <: AbstractVector{T}

One-dimensional array with elements of type T with a given sampling distance. Alias for SpatialArray{T,1}.

Transfer Functions

TransferFunctions.TransferFunction — Type

TransferFunction

Linear Transfer Functions

TransferFunctions.LinearTransferFunction — Type

LinearTransferFunction <: TransferFunction

TransferFunctions.conv — Function

conv(t::LinearTransferFunction, ::SpatialArray{<:Real,2})

TransferFunctions.deconv — Function

deconv(t::LinearTransferFunction, ::SpatialArray{<:Real,2})

Optical Transfer Functions

TransferFunctions.OpticalTransferFunction — Type

OpticalTransferFunction <: LinearTransferFunction

Implementation

To create a new Optical transfer function (OTF) A <: OpticalTransferFunction, you must define the attenuation at a given frequency coordinate attenuation(otf::A, kx::Frequency, ky::Frequency).

TransferFunctions.OTFArray — Type

OTFArray{T<:Number} <: OpticalTransferFunction

Fields

data::AbstractMatrix{T}

Δxy::PixelSize{2}

center::Coordinate{2}

extend::ExtensionMethod

TransferFunctions.RadialOTF — Type

RadialOTF <: OpticalTransferFunction

If the pupil function of the system is symmetric, the OTF as well as the PSF are radially symmetric which can be used to optimize the calculations.

Implementation

attenuation(model::A, fr::Frequency)

cutoff(model::A, [a=0]) returning the largest frequency f such that attenuation(model, f) >= a if a > 0 and

attenuation(model, f) > 0 if a = 0.

TransferFunctions.CircularPupilOTF — Type

CircularPupilOTF(ρ₀::Real)

Ideal (aberration free) OTF of a diffraction limited imaging system with incoherent light with the cutoff-frequency ρ₀

TransferFunctions.otf — Function

otf(tf::OpticalTransferFunction, Δxy::PixelSize{2}, wh::Dims{2})

TransferFunctions.insupport — Function

insupport(otf::OpticalTransferFunction, fx::Frequency, fy::Frequency)

TransferFunctions.cutoff — Function

cutoff(::RadialOTF, a=0.0)

TransferFunctions.attenuation — Function

attenuation(otf::OpticalTransferFunction, ::Frequency, ::Frequency)

Point Spread Functions

TransferFunctions.PointSpreadFunction — Type

PointSpreadFunction <: LinearTransferFunction

Implementation

To create a new Point spread function (PSF) A <: PointSpreadFunction, you must define the intensity at a given length coordinate intensity(psf::A, x::Length, y::Length).

TransferFunctions.MeasuredPSF — Type

MeasuredPSF <: PointSpreadFunction

TransferFunctions.PSFArray — Type

PSFArray{T<:Real} <: PointSpreadFunction

A measurement of a PSF from the acquired image(s).

Fields

data::AbstractMatrix{T}

Δ::PixelSize{2}

center::Coordinate{2}

extend::ExtensionMethod

TransferFunctions.ModelPSF — Type

ModelPSF <: PointSpreadFunction

TransferFunctions.RadialPSF — Type

RadialPSF <: ModelPSF

TransferFunctions.BornWolf — Type

BornWolf{T<:Real} <: RadialPSF

Born & Wolf model point spread function

Fields

+ `λ::Length{T}`: emission wavelength 
+ `n::T`: index of refraction of the immersion medium
+ `NA::T`: numerical aperture of the objective

TransferFunctions.GibsonLanni — Type

GibsonLanni{T<:Real} <: RadialPSF

Gibson & Lanni model point spread function

Fields

+ `λ::Length{T}`: emission wavelength
+ `NA::T`: numerical aperture of the objective
+ `n_i::T`: refractive index of the immersion medium
+ `n_s::T`: refractive index of the sample
+ `n_g::T`: refractive index of the coverslip
+ `t_i::Length{T}`: working distance of the objective
+ `t_g::Length{T}`: coverslip thickness

TransferFunctions.psf — Function

psf(tf::PointSpreadFunction, Δ::PixelSize{2}, wh::Dims{2})

TransferFunctions.intensity — Function

intensity(psf::PointSpreadFunction, ::Length, ::Length)

TransferFunctions.fit — Function

fit(::ModelPSF, ::SpatialArray)

Estimation

TransferFunctions.Estimation.bead — Function

bead([T=Float64], d, Δ; <kwargs>)

Generate a model of a fluorescent microsphere (AKA calibration bead) with diameter d and pixel-size Δ.

Arguments

α=0u"nm^-1: evanescent wave attenuation constant in the TIR-FM microscopy modulation. An acquisition that is not TIR-FM is equivalent to setting α=0u"nm^-1".

pixel_grid=10: aliasing of the pixels is done by averaging values at a larger grid. pixel_grid sets the dimensions equivalent to each pixel in the grid

intesity=1.0: peak intensity value, i.e. the theoretical value at the center of the bead

position=(0.0, 0.0): set the subpixel position of the peak (center) of the bead within the generated array

Examples

julia> Estimation.bead(100u"nm", 30.5u"nm");

julia> Estimation.bead(95u"nm", (30.5u"nm", 25u"nm"));

julia> Estimation.bead(Float32, 0.1u"μm", 30.5u"nm");

julia> Estimation.bead(0.1u"μm", 30.5u"nm"; α=0.01u"nm^-1")
5×5 OffsetArray(::Matrix{Float64}, -2:2, -2:2) with eltype Float64 with indices -2:2×-2:2:
 0.0        0.0       0.0705075  0.0       0.0
 0.0        0.606779  0.892914   0.606779  0.0
 0.0705075  0.892914  1.0        0.892914  0.0705075
 0.0        0.606779  0.892914   0.606779  0.0
 0.0        0.0       0.0705075  0.0       0.0

julia> Estimation.bead(0.1u"μm", 50.5u"nm"; position=(-0.3,-0.2))
3×3 OffsetArray(::Matrix{Float64}, -1:1, -1:1) with eltype Float64 with indices -1:1×-1:1:
 0.333333  0.737374  0.0808081
 0.59596   1.0       0.20202
 0.020202  0.141414  0.0

julia> Estimation.bead(0.1u"μm", 50.5u"nm"; intensity=0.5)
3×3 OffsetArray(::Matrix{Float64}, -1:1, -1:1) with eltype Float64 with indices -1:1×-1:1:
 0.03  0.23  0.03
 0.23  0.5   0.23
 0.03  0.23  0.03

Apodization

TransferFunctions.Apodization — Module

TransferFunctions.Apodization module defines apodization functions of various formats and utility methods that are related to apodization such as taperedges and apodize.

Apodization types Blackman, ExactBlackman, Connes, Cosine, Gaussian, Hamming, Welch, BlackmanNuttall, PowerCosine, Triangular, Nuttall, SineSum, BlackmanHarris, FlatTop and Hann

Exports taperedges, apodize

TransferFunctions.Apodization.apodize — Function

apodize(apo::ApodizationFunction, A, cutoff::Real, width::Real)

Apply apodization to the input array A.

Arguments

apo::ApodizationFunction: The type of apodization to apply.

cutoff::Real: The cutoff frequency for the apodization.

width::Real: The width of the transition region for the apodization.

TransferFunctions.Apodization.taperedges — Function

taperedges([apo=Cosine()], A, w, [border=:replicate]; dims=:)
taperedges([apo], A, (w1,w2...,wn), [border]; dims =1:n)
taperedges([apo], A, ((w1a,...,wna), (w1b,...,wnb)), [border]; dims=1:n)

Taper edges of width w of array A using apo::ApodizationFunction.

Widths can be same for all dimensions and directions using w::Int, specified separately for each of the n dimensions (w1,w2...,wn) which means that the 'start' and 'end' padding will be the same. The third option is to specify the padding fully separately for the 'start' and 'end' of n dimensions ((w1a,...,wna), (w1b,...,wnb)). In the last two cases, the dimensions are specified by dims and by default are taken as the first n dimensions. Border can be :replicate, :circular, :symmetric, :reflect or Fill(v).

See also BorderArray, ApodizationFunction

TransferFunctions.Apodization.ApodizationFunction — Type

ApodizationFunction

Abstract type for apodization functions.

TransferFunctions.Apodization.Blackman — Type

Blackman{α,T} == SineSum{3,((1 - α) / 2, 1 / 2, α / 2),T} <: ApodizationFunction

Default is α = 0.16.

See also ExactBlackman

TransferFunctions.Apodization.ExactBlackman — Type

ExactBlackman == Blackman{683 // 4652}

TransferFunctions.Apodization.Connes — Type

Connes <: ApodizationFunction

TransferFunctions.Apodization.Cosine — Type

Cosine == PowerCosine{1} <: ApodizationFunction

zero-phase function: $w₀(r) = cos(πr/2)$

instrument function: $I(k) = 4cos(2k)/(π(1 - 16k²))$

TransferFunctions.Apodization.Gaussian — Type

Gaussian{T<:Real} <: ApodizationFunction

Gaussian apodization function with a given standard deviation σ.

zero-phase function: $w₀(r) = e^{-r²/2σ²}$

TransferFunctions.Apodization.Hamming — Type

Hamming == SineSum{2,(25//46, 21//46)} <: ApodizationFunction

TransferFunctions.Apodization.Welch — Type

Welch <: ApodizationFunction

zero-phase function: $w₀(r) = 1 - r²$

TransferFunctions.Apodization.BlackmanNuttall — Type

BlackmanNuttall == SineSum{4,(0.3635819, 0.4891775, 0.1365995, 0.0106411) } <: ApodizationFunction

TransferFunctions.Apodization.PowerCosine — Type

PowerCosine{α<:Real} <: ApodizationFunction

zero-phase function: $w₀(r) = cos(πr/2)^α$

Instances: Cosine and Hann

TransferFunctions.Apodization.Triangular — Type

Triangular <: ApodizationFunction

Formulas:

zero-phase function: $w₀(r) = 1-|r|$

instrument function: $I(k) = sinc²(π k)$

TransferFunctions.Apodization.Nuttall — Type

Nuttall} == SineSum{4,(0.355768, 0.487396, 0.144232, 0.012604)} <: ApodizationFunction

TransferFunctions.Apodization.SineSum — Type

SineSum{N,Cs,T} <: ApodizationFunction

Sum of N sines apodization function with output type of T

zero-phase function: $w₀(r) = ∑ᴺₖ₌₀ (-1)ᵏ Cs[k] cos(πk(r + 1))$

Instances: Hamming, Nuttall, BlackmanNuttall, BlackmanHarris, FlatTop, Blackman and ExactBlackman

TransferFunctions.Apodization.BlackmanHarris — Type

BlackmanHarris == SineSum{4,(0.35875, 0.48829, 0.14128, 0.01168)} <: ApodizationFunction

TransferFunctions.Apodization.FlatTop — Type

FlatTop == SineSum{5,(0.21557895, 0.41663158, 0.277263158, 0.083578947, 0.006947365)} <: ApodizationFunction

MATLAB variant of the flat-top filter

TransferFunctions.Apodization.Hann — Type

Hann == PowerCosine{2} <: ApodizationFunction

Formulas:

zero-phase function: $w₀(r) = cos²(πr/2)$

Internals

Functions

TransferFunctions.roundupcenter — Function

roundupcenter(A)

Calculate center index of A rounded-up (i.e. fft center).

Examples

julia> roundupcenter(ones(15,15))
CartesianIndex(8, 8)

julia> roundupcenter((16,16))
CartesianIndex(9, 9)

TransferFunctions.rounddowncenter — Function

rounddowncenter(A)

Calculate center index of A rounded-down (i.e. ifft center).

Examples

julia> rounddowncenter(ones(15,15))
CartesianIndex(8, 8)

julia> rounddowncenter((16,16))
CartesianIndex(8, 8)

TransferFunctions.exactcenter — Function

 exactcenter(A)

Calculate the exact center coordinate of A.

Examples

julia> exactcenter(ones(15,15))
(8.0, 8.0)

julia> exactcenter((16,16))
(8.5, 8.5)

TransferFunctions.aroundorigin — Function

aroundorigin(s::Size, origin=(0,0,...))

Return the offset axes of an array of size s around the origin.

TransferFunctions.contained — Function

contained(A, loc::Coordinate{N})

Return true if the coordinate loc is contained in the axes of array A.

TransferFunctions.interior — Function

interior(inds::Indices{N}, kern::Indices{N})

Return 'valid' indices for convolution of an array with indices inds with a kernel having the indices kern.

TransferFunctions.@__FUNCTION__ — Macro

@__FUNCTION__

Return current function name. For debugging and error message purposes.

Type Aliases

TransferFunctions.Size — Type

Size{N}

N-dimensional size of an array. Alias for NTuple{N,Int}

TransferFunctions.Coordinate — Type

Coordinate{N,T}

N-dimensional coordinate. Alias for NTuple{N,T<:Real}

TransferFunctions.PixelSize — Type

PixelSize{N}

N-dimensional pixel size. Alias for NTuple{N,Length}

TransferFunctions.SBitVector — Type

SBitVector{N}

N-dimensional static vector with elements of type Bool. Alias for SVector{N,Bool}