OTF Models

TransferFunctions.ModelOTF — Type

Implementation

To create a new OTF model A <: ModelOTF, you need to implement the functions:

The transfer coefficient at a give frequency:

attenuation(model::A, kx::Frequency, ky::Frequency) for a non-symmetric OTF, or

attenuation(model::A, kᵣ::Frequency) for a symmetric OTF, where you also have to call @traitimpl RadiallySymmetric{A}

You should consider implementing the following optional methods:

preferred_type(::Type{<:ModelOTF}) method, which returns the natural type that is returned from calling the model. It should be either <:Complex or <:Real.

For a symmetric OTF, you should consider implementing the optional methods:

cutoff(model::A; a) which returns the largest frequency f::Frequency such that attenuation(model, f) >= a if a > 0 and attenuation(model, f) > 0 if a == 0

TransferFunctions.IdealOTFwithCurvature — Type

IdealCircularPupilOTF(ρ₀::Real)

Ideal (aberration free) OTF of a diffraction limited imaging system with incoherent light with the cutoff-frequency ρ₀

The OTF is derived from the diffraction caused by the exit pupil of the system and disregards the effect of the entrance pupil... thus assumes no reshaping of the wavefronts in the optical system. The exit pupil, being located in the optical system just before the light reaches the image plane, has a greater effect on the optical system OTF.

See also IdealCircularPupilPSF (TODO)

Examples

- TODO

Extended help

The OTF can be written in as a function of the cutoff frequency $ρ₀$ [1]

\[ ℋ(ρ) = \begin{cases} (2/π) \left\{ \arccos(ρ/2ρ₀) - (ρ/2ρ₀)\sqrt{1 - (ρ/2ρ₀)²} \right\} & \text{ for } ρ ≤ 2ρ₀ \\ 0 & \text{ otherwise}. \end{cases} \]

The cutoff frequency can be written in terms of the wavelength $λ$, distance between entrance pupil and the image plane $f₂$ and the circular pupil radius $w$ as

\[ \rho_0 = w/(λ f₂).\]

PSF Models

TransferFunctions.BornWolf — Type

Born & Wolf model of the transfer function for a circular aperture.

Parameters

λ::Unitful.Length: (emission) wavelength

nᵢ::Number: index of refraction of the immersion medium

NA::Number: numerical aperture

Extended help

The Born & Wolf model is a scalar diffraction model derived for perfect systems. It assumes that the only aberration of the system is due to defocus. Modern microscope objectives are designed to provide optimal imaging conditions for sources located directly on the coverslip, in which case the Born & Wolf model is applicable (if the coverslip and immersion is used as designed). The model disregards spherical and higher order aberrations that are due to the source of illumination being shifted from the coverslip boundary.

TransferFunctions.GibsonLanni — Type

Gibson & Lanni model of the transfer function for a circular aperture.

Parameters

Extended help

The Gibson & Lanni model is assumes that, disregarding defocus, all observed aberrations are generated by factors external to the objective (i.e. originating in the sample, coverslip and immersion medium combination). These aberrations can be characterized by the optical path difference between a ray in a perfect system (see BornWolf) and a ray under experimental conditions.

1

Frequency Analysis of Optical Imaging Systems. In Introduction to Fourier optics; Roberts & Co: Englewood, Colo, 2005; pp. 127–172 ISBN 978-0-9747077-2-3.