TransferFunctions.TransferFunction — Type

An N-dimensional transfer function

TransferFunctions.MeasuredPSF — Type

MeasuredPSF holds an array of measured data with information about the dimensions of the measurement. This allows you to make conversions to other representations of a transfer function such as an OTF and to use interpolations to sample the PSF at arbitrary locations not included in the initial measurement (in general to use the measurements similarly to how one could use a model). This can be useful for using the same PSF measurement for an acquisition with a different pixelsize or in super resolution applications.

A measurement can be done for example using an image of subresolution microspheres.

TransferFunctions.MeasuredOTF — Type

MeasuredOTF holds an array of measured data with information about the dimensions of the measurement. This allows you to make conversions to other representations of a transfer function such as a PSF and to use interpolations to sample the OTF at arbitrary locations not included in the initial measurement (in general to use the measurements similarly to how one could use a model). This can be useful for using the same OTF measurement for an acquisition with a different pixelsize or in super resolution applications.

TransferFunctions.SampledTransferFunction — Type

TODO

Functions

TransferFunctions.psf — Function

Intensity point spread function i.e. the intensity ratio and phase shift of the sample intensity density

TransferFunctions.ipsf — Function

Intensity point spread function

TransferFunctions.apsf — Function

Amplitude point spread function

TransferFunctions.otf — Function

optical transfer function

TransferFunctions.pupil — Function

pupil function

TransferFunctions.ptf — Function

phase transfer function

TransferFunctions.mtf — Function

modulation transfer function

Utility Functions and Traits

TransferFunctions.roundupcenter — Function

roundupcenter(arr::AbstractArray)

Calculate center index of arr rounded-up (i.e. fft center)

Examples

julia> roundupcenter(ones(15,15))
(8, 8)

julia> roundupcenter(ones(16,16))
(9, 9)

julia> roundupcenter(ones(15,15,15))
(8, 8, 8)

TransferFunctions.exactcenter — Function

 exactcenter(a::AbstractArray)

Calculate the exact center index of a

Examples

julia> exactcenter(ones(15,15))
(8.0, 8.0)

julia> exactcenter(ones(16,16))
(8.5, 8.5)

julia> exactcenter(ones(15,15,15))
(8.0, 8.0, 8.0)

TransferFunctions.RadiallySymmetric — Type

If the pupil function of the system is symmetric, the OTF as well as the PSF are radially symmteric which can be used to optimize the calculations.

Apodization

TransferFunctions.apodize — Function

apodize(apo::Apodization, A::AbstractArray{<:Number,N}, cutoff::Real, width::Real) where {N}

Apply apodization to the input array A.

Arguments

apo::Apodization: The type of apodization to apply.

A::AbstractArray{<:Number,N}: The input array to be apodized.

cutoff::Real: The cutoff frequency for the apodization.

width::Real: The width of the transition region for the apodization.

TransferFunctions.taperedges — Function

taperedges(apo, A, width::Int, [dims])
taperedges(apo, A, (w1,...,wM), [(d1,...,dM)])
taperedges(apo, A, ((w1_start,...,wM_start), (w1_end,...,wM_end)), [(d1,...,dM)])

Taper the edges of the input array A using the apodization function apo::Apodization, along dims::Tuple

apo::Apodization: apodization function to apply.

A::AbstractArray{<:Number,N}: The input array to taper.

width::NTuple{M,Int} or Int or Tuple{NTuple{M,Int},NTuple{M,Int}}: The width of tapering at the edges.

dims::NTuple{M,Int}: The dimensions along which to taper the edges.

Apodization Functions

TransferFunctions.Bartlett — Type

Bartlett <: Apodization

$A(x) = 1-|x|/a$

$I(k) = a \mathop{sinc}²(π k a)$

Examples:

Bartlett(5) <: Apodization

TransferFunctions.Blackman — Type

Blackman(a::Real) <: Apodization

$A(x) = 0.42 + 0.5 \cos(\pi x/a) + 0.08 \cos(2\pi x/a)$

$I(k) = \frac{a \mathop{sinc}(2\pi k a) (0.84 - 0.36a^2k^2)}{(1 - a^2k^2)(1 - 4a^2k^2)}$

where $a$ is the apodization parameter.

TransferFunctions.Connes — Type

Connes(a::Real) <: Apodization

TransferFunctions.Cosine — Type

Cosine(a::Real) <: Apodization

Cosine apodization function with a given amplitude factor a.

Formulas:

Apodization Function: $\cos\left(\pi \frac{x}{2a}\right)$

Instrument Function: $\frac{4a\cos\left(2ak\right)}{\pi\left(1 - 16a^2 k^2\right)}$

TransferFunctions.Gaussian — Type

Gaussian(σ::Real) <: Apodization

Gaussian apodization function with a given standard deviation σ.

Formulas: - Apodization Function: $e^{-\frac{x^2}{2\sigma^2}}$ - Instrument Function: Not implemented (Low Priority)

TransferFunctions.Hamming — Type

Hamming(a::Real) <: Apodization

TransferFunctions.Hanning — Type

Hanning(a::Real) <: Apodization

TransferFunctions.Welch — Type

Welch(a::Real) <: Apodization

To Deprecate

TransferFunctions.padtosize — Function

padtosize(a::AbstractArray{T,N}, size...; fourier=false, padvalue=0)
padtosize(a::AbstractArray{T,N}, size; fourier=false, padvalue=0)

Pad a to a size

Padding is done in a way which ensures that if a is symmetric and decreases to zero on the edges (!!) such that eltype(fft(a)) <: Real, then, if possible, output is symmetric under the DFT and elyptype(fft(a_padded)) <: Real). If a is in the Fourier domain, then fourier should de set to true.

Arguments:

size: size of the output array. If size isa Integer then size along all the dimensions is size

fourier: a is in the Fourier domain (default: false).

Examples:

julia> padtosize(reshape(1:4, 2,2), 3, 3)
3×3 Matrix{Int64}:
 0  0  0
 0  1  3
 0  2  4

julia> padtosize(reshape(1:4, 2,2), 4, 4)
4×4 Matrix{Int64}:
 0  0  0  0
 0  1  3  0
 0  2  4  0
 0  0  0  0

julia> padtosize(reshape(1:4, 2,2), 3, 3; fourier=true)
3×3 Matrix{Int64}:
 4  2  0
 3  1  0
 0  0  0

julia> padtosize(reshape(1:4, 2,2), 4, 4; fourier=true)
4×4 Matrix{Int64}:
 1  0  0  3
 0  0  0  0
 0  0  0  0
 2  0  0  4

julia> ones(1,1) |> fft
1×1 Matrix{ComplexF64}:
 1.0 + 0.0im

julia> padtosize(ones(1,1), 2,2) |> fft
2×2 Matrix{ComplexF64}:
  1.0+0.0im  -1.0+0.0im
 -1.0+0.0im   1.0+0.0im

julia> ones(2,2) |> fft
2×2 Matrix{ComplexF64}:
 4.0+0.0im  0.0+0.0im
 0.0+0.0im  0.0+0.0im

julia> padtosize(ones(2,2), 3, 3) |> fft # can be made symmetric
3×3 Matrix{ComplexF64}:
  4.0+0.0im  -2.0+0.0im  -2.0+0.0im
 -2.0+0.0im   1.0+0.0im   1.0+0.0im
 -2.0+0.0im   1.0+0.0im   1.0+0.0im

julia> padtosize(ones(2,2), 4, 4) |> fft # cannot be symmetric
4×4 Matrix{ComplexF64}:
  4.0+0.0im  -2.0-2.0im  0.0+0.0im  -2.0+2.0im
 -2.0-2.0im   0.0+2.0im  0.0+0.0im   2.0+0.0im
  0.0+0.0im   0.0+0.0im  0.0+0.0im   0.0+0.0im
 -2.0+2.0im   2.0+0.0im  0.0+0.0im   0.0-2.0im

julia> @assert all(-10^-5 .< imag(fft(padtosize(ones(2,2), 5, 5))) .< 10^-5) # .== 0 (numerical errors)