TransferFunctions.jl

Transfer functions of optical systems their models, estimation, conversion, visualization, analysis, simulation and application to images.

TransferFunctions.TransferFunctions — Module

Package for models, estimation, sampling and deconvolution with microscopy transfer functions.