TransferFunctions

Documentation for TransferFunctions.

TransferFunctions.TransferFunctions

TransferFunctions.BornWolf

TransferFunctions.GibsonLanni

TransferFunctions.MeasuredPSF

TransferFunctions.MeasuredTransferFunction

TransferFunctions.ModelTransferFunction

TransferFunctions.RadiallySymmetric

TransferFunctions.apsf

TransferFunctions.apsf

TransferFunctions.ipsf

TransferFunctions.ipsf

TransferFunctions.mtf

TransferFunctions.otf

TransferFunctions.psf

TransferFunctions.psf

TransferFunctions.psf

TransferFunctions.psf

TransferFunctions.ptf

TransferFunctions.pupil

TransferFunctions.TransferFunctions — Module

This module implements diffraction transfer function models that are widely used in microscopy. The base type for any transfer function is TransferFunction and further disambiguates to MeasuredTransferFunction and ModelTransferFunction.

Warning

All of the models assume incoherent illumination sources (i.e. where the radiation from the source is inocoherent). This is an approximation for most optical setups, but one that is used almost universally (e.g. in fluorescence microscopy).

TransferFunctions.BornWolf — Type

Born & Wolf model of the transfer function for a circular aperture.

Parameters

λ::Unitful.Length: (emission) wavelength

nᵢ::Number: index of refraction of the immersion medium

NA::Number: numerical aperture

Extended help

The Born & Wolf model is a scalar diffraction model derived for perfect systems. It assumes that the only aberration of the system is due to defocus. Modern microscope objectives are designed to provide optimal imaging conditions for sources located directly on the coverslip, in which case the Born & Wolf model is applicable (if the coverslip and immersion is used as designed). The model disregards spherical and higher order aberrations that are due to the source of illumination being shifted from the coverslip boundary.

TransferFunctions.GibsonLanni — Type

Gibson & Lanni model of the transfer function for a circular aperture.

Parameters

Extended help

The Gibson & Lanni model is assumes that, disregarding defocus, all observed aberrations are generated by factors external to the objective (i.e. originating in the sample, coverslip and immersion medium combination). These aberrations can be characterized by the optical path difference between a ray in a perfect system (see BornWolf) and a ray under experimental conditions.

TransferFunctions.MeasuredPSF — Type

MeasuredPSF holds an array of measured data with information about the dimensions of the measurement. This allows you to make conversions to other representations of a transfer function such as an OTF and to use interpolations to sample the PSF at arbitrary locations not included in the initial measurement (in general to use the measurements similarly to how one could use a model). This can be useful for using the same PSF measurement for an acquisition with a different pixelsize or in super resolution applications.

A measurement can be done for example using an image of subresolution microspheres.

TransferFunctions.MeasuredTransferFunction — Type

An abstract type for the measurement of the transfer function of an optical system. It can be either a PSF measurement MeasuredPSF or a MeasuredOTF.

TransferFunctions.ModelTransferFunction — Type

An abstract type for any transfer function that is based on a physical model of an optical system. Contrary to a MeasuredTransferFunction, a ModelTransferFunction must be quantifiable at any point (in either spatial or frequency domain). The model can be of an OTF (ClosedFormOTFModel <: ModelTransferFunction) or a PSF (ClosedFormPSFModel <: ModelTransferFunction) or a PupilFunction.

TransferFunctions.RadiallySymmetric — Type

If the pupil function of the system is symmetric, the OTF as well as the PSF are radially symmteric which can be used to optimize the calculations

TransferFunctions.apsf — Function

Amplitude point spread function

TransferFunctions.apsf — Method

apsf(tf::TransferFunction, wh::Tuple{Integer,Integer}, Δxy::Tuple{Length,Length})::OffsetMatrix{<:Real}
apsf(tf::TransferFunction, wh::Tuple{Integer,Integer}, Δxy::Length)::OffsetMatrix{<:Real}

TransferFunctions.ipsf — Function

Intensity point spread function

TransferFunctions.ipsf — Method

ipsf(tf::TransferFunction, wh::Tuple{Integer,Integer}, Δxy::Tuple{Length,Length})::OffsetMatrix{<:Real}
ipsf(tf::TransferFunction, wh::Tuple{Integer,Integer}, Δxy::Length)::OffsetMatrix{<:Real}

TransferFunctions.mtf — Method

modulation transfer function

TransferFunctions.otf — Function

optical transfer function

TransferFunctions.psf — Function

Intensity point spread function i.e. the intensity ratio and phase shift of the sample intensity density

TransferFunctions.psf — Method

psf(tf::MeasuredPSF, wh::Tuple{Integer,Integer} [,Δxy::Tuple{Length,Length}]; intp, extp)::OffsetMatrix

Returns a centered PSF interpolated from the measurement.

Parameters

intp = BSpline(Cubic(Flat(OnGrid()))): interpolation method (default: cubic interpolation with flat slopes on the edge knots)

extp = zero(eltype(tf.data)): extrapolation method

zero(eltype(tf.data)) constant extrapolation outside of the bounds (ideal for background subtracted PSF)

Flat() constant slope beyond the bounds (gives the last knot value and is equivalent to 0 for a background subtracted PSF)

δ = wh .÷ 2: output center of the PSF (default: center of the output size array rounded down)

Extended Help

Implementation Details

Interpolation is done using scaled BSplines.

TransferFunctions.psf — Method

psf(tf::TransferFunction, wh::Tuple{Integer,Integer}, Δxy::Tuple{Length,Length})::OffsetMatrix
psf(tf::TransferFunction, wh::Tuple{Integer,Integer}, Δxy::Length)::OffsetMatrix

Generate the psf with size wh with Δxy being the distance between the samples in the x and y dimensions

julia> tf = BornWolf(488u"nm", 1.7, 1.7)
BornWolf{Float64}(488.0 nm, 1.7, 1.7)

tf = BornWolf(488u"nm", 1.7, 1.7) # hide
psf(tf, (11,11), 60u"nm")
psf(tf, (5,5), (40u"nm", 50u"nm")) # different pixelsizes in x and y direction

TransferFunctions.psf — Method

psf(tf::TransferFunction, x::Length, y::Length)

Sample the PSF of the transfer function model/data at the point (x,y)

julia> tf = BornWolf(488u"nm", 1.4, 1.7)
BornWolf{Float64}(488.0 nm, 1.4, 1.7)

tf = BornWolf(488u"nm", 1.4, 1.7) # hide
psf(tf, 0u"nm", 5u"nm")
psf.(tf, 0u"nm", -400u"nm":100u"nm":400u"nm")

TransferFunctions.ptf — Method

phase transfer function

TransferFunctions.pupil — Method

pupil function