Optical Transfer Functions

TransferFunctions.OTFArray — Type

OTFArray{T<:Number,N} <: OpticalTransferFunction{N}

Fields

data::AbstractMatrix{T}

center::Coordinate{2}

TransferFunctions.RadialOTF — Type

RadialOTF{N} <: OpticalTransferFunction{N}

If the pupil function of the system is symmetric, the OTF as well as the PSF are radially symmetric which can be used to optimize the calculations.

Implementation

attenuation(model::A, fr::Frequency)

cutoff(model::A, [a=0]) returning the largest frequency f such that attenuation(model, f) >= a if a > 0 and

attenuation(model, f) > 0 if a = 0.

TransferFunctions.otf — Function

otf(tf::OpticalTransferFunction, Δxy::PixelSize{2}, wh::Dims{2})

TransferFunctions.insupport — Function

insupport(otf::OpticalTransferFunction, fx::Frequency, fy::Frequency)

TransferFunctions.cutoff — Function

cutoff(::RadialOTF, a=0.0)

TransferFunctions.attenuation — Function

attenuation(otf::OpticalTransferFunction, ::Frequency, ::Frequency)

Point Spread Functions

TransferFunctions.psf — Function

psf(tf::PointSpreadFunction, Δ, wh::Dims; normalize=true)

Generate a PSF array size wh for the model tf with the pixel size Δ.

psf(tf::PointSpreadFunction{2}, Δ, ε::Real; <kwargs>)

Sample the PSF tf with a pixel size Δ over a window such that the energy error is less than ε.

kwargs are passed to the final psf function.

TransferFunctions.fit — Function

fit(::PSFModel, ::SpatialArray)

TransferFunctions.FWHM — Function

FWHM(psf::PointSpreadFunction)

Find the FWHM of the PSF psf in the x and y directions.

See also HWHM

julia> tf = IsotropicGaussian(488u"nm", 1.4);

julia> TF.FWHM(tf)
(179.33894905055288 nm, 179.33894905055288 nm, 93.61250264937092 nm)

TransferFunctions.HWHM — Function

HWHM(psf::PointSpreadFunction)

Find the HWHM of the PSF psf in all the axes directions.

See also FWHM

julia> tf = AiryDisc(λ=488u"nm", NA=1.4);

julia> TransferFunctions.HWHM(tf)
((-89.66947452527637 nm, 89.66947452527637 nm), (-89.66947452527637 nm, 89.66947452527637 nm), (-147.04617530370933 nm, 147.04617530370933 nm))

Non-linear Transfer Functions

TransferFunctions.ImpulseResponseMapping — Type

ImpulseResponseMapping{2} <: LinearTransferFunction{2}

An impulse response mapping is a description of a transfer function that prescribes an impulse response to every point in the object plane. This may be a point spread function or some other function that prescribes the response to the object plane point in the image plane.

An transfer defined by an impulse response mapping is a linear system, but is not translation invariant hence does not belong to the class of linear transfer functions.

Internal Utility Functions and Types

Functions

TransferFunctions.roundcenter — Function

roundcenter(r::RoundingMode, A)

Calculate the center index of A (::AbstractArray/::Indices/::Size/::AbstractUnitRange) rounded with the mode r::RoundingMode.

See also roundupcenter, rounddowncenter, exactcenter

TransferFunctions.roundupcenter — Function

roundupcenter(A)

Calculate center index of A rounded-up (i.e. fft center).

Examples

julia> TF.roundupcenter(ones(15,15))
CartesianIndex(8, 8)

julia> TF.roundupcenter((16,16))
CartesianIndex(9, 9)

TransferFunctions.rounddowncenter — Function

rounddowncenter(A)

Calculate center index of A rounded-down (i.e. ifft center).

Examples

julia> TF.rounddowncenter(ones(15,15))
CartesianIndex(8, 8)

julia> TF.rounddowncenter((16,16))
CartesianIndex(8, 8)

TransferFunctions.exactcenter — Function

 exactcenter(A)

Calculate the exact center coordinate of A.

Examples

julia> TF.exactcenter(ones(15,15))
(8.0, 8.0)

julia> TF.exactcenter((16,16))
(8.5, 8.5)

TransferFunctions.aroundorigin — Function

aroundorigin(s::Size, origin=(0,0,...))

Return the offset axes of an array of size s around the origin.

TransferFunctions.contained — Function

contained(A, loc::Coordinate{N})

Return true if the coordinate loc is contained in the axes of array A.

TransferFunctions.interior — Function

interior(inds::Indices{N}, kern::Indices{N})

Return 'valid' indices for convolution of an array with indices inds with a kernel having the indices kern.

TransferFunctions.fftfreqs — Function

fftfreqs(sz::Size{2}, Δ::PixelSize{2})
fftfreqs(sz::Size{2}, Δ::Length)

Generate a frequency grid for 2D FFTs with the pixel size Δ

julia> x_f, y_f = TF.fftfreqs((5,5), 50u"nm");

julia> x_f
5×5 Matrix{Quantity{Float64, 𝐋^-1, Unitful.FreeUnits{(nm^-1,), 𝐋^-1, nothing}}}:
  0.0 nm^-1     0.0 nm^-1     0.0 nm^-1     0.0 nm^-1     0.0 nm^-1
  0.004 nm^-1   0.004 nm^-1   0.004 nm^-1   0.004 nm^-1   0.004 nm^-1
  0.008 nm^-1   0.008 nm^-1   0.008 nm^-1   0.008 nm^-1   0.008 nm^-1
 -0.008 nm^-1  -0.008 nm^-1  -0.008 nm^-1  -0.008 nm^-1  -0.008 nm^-1
 -0.004 nm^-1  -0.004 nm^-1  -0.004 nm^-1  -0.004 nm^-1  -0.004 nm^-1

julia> y_f
5×5 Matrix{Quantity{Float64, 𝐋^-1, Unitful.FreeUnits{(nm^-1,), 𝐋^-1, nothing}}}:
 0.0 nm^-1  0.004 nm^-1  0.008 nm^-1  -0.008 nm^-1  -0.004 nm^-1
 0.0 nm^-1  0.004 nm^-1  0.008 nm^-1  -0.008 nm^-1  -0.004 nm^-1
 0.0 nm^-1  0.004 nm^-1  0.008 nm^-1  -0.008 nm^-1  -0.004 nm^-1
 0.0 nm^-1  0.004 nm^-1  0.008 nm^-1  -0.008 nm^-1  -0.004 nm^-1
 0.0 nm^-1  0.004 nm^-1  0.008 nm^-1  -0.008 nm^-1  -0.004 nm^-1

TransferFunctions.posgrid — Function

posgrid(axes::Indices, Δ)
posgrid(sz::Size, Δ; center=roundupcenter(sz))

Generate a position grid for 2D sampled images with the pixel size Δ.

If a size sz is passed generate a position grid with the given size and the center in center. If Δ is a tuple of Lengths then the elements are used for the sampling in the respective dimensions. If Δ is a single length, then it is used for all dimensions.

julia> x_p, y_p = TF.posgrid((-3:5, -1:3), (50u"nm", 20u"nm"));

julia> x_p
9×5 Matrix{Quantity{Float64, 𝐋, Unitful.FreeUnits{(nm,), 𝐋, nothing}}}:
 -150.0 nm  -150.0 nm  -150.0 nm  -150.0 nm  -150.0 nm
 -100.0 nm  -100.0 nm  -100.0 nm  -100.0 nm  -100.0 nm
  -50.0 nm   -50.0 nm   -50.0 nm   -50.0 nm   -50.0 nm
    0.0 nm     0.0 nm     0.0 nm     0.0 nm     0.0 nm
   50.0 nm    50.0 nm    50.0 nm    50.0 nm    50.0 nm
  100.0 nm   100.0 nm   100.0 nm   100.0 nm   100.0 nm
  150.0 nm   150.0 nm   150.0 nm   150.0 nm   150.0 nm
  200.0 nm   200.0 nm   200.0 nm   200.0 nm   200.0 nm
  250.0 nm   250.0 nm   250.0 nm   250.0 nm   250.0 nm

julia> y_p
9×5 Matrix{Quantity{Float64, 𝐋, Unitful.FreeUnits{(nm,), 𝐋, nothing}}}:
 -20.0 nm  0.0 nm  20.0 nm  40.0 nm  60.0 nm
 -20.0 nm  0.0 nm  20.0 nm  40.0 nm  60.0 nm
 -20.0 nm  0.0 nm  20.0 nm  40.0 nm  60.0 nm
 -20.0 nm  0.0 nm  20.0 nm  40.0 nm  60.0 nm
 -20.0 nm  0.0 nm  20.0 nm  40.0 nm  60.0 nm
 -20.0 nm  0.0 nm  20.0 nm  40.0 nm  60.0 nm
 -20.0 nm  0.0 nm  20.0 nm  40.0 nm  60.0 nm
 -20.0 nm  0.0 nm  20.0 nm  40.0 nm  60.0 nm
 -20.0 nm  0.0 nm  20.0 nm  40.0 nm  60.0 nm

julia> TF.posgrid((3,3,3), 50u"nm")[3]
3×3×3 Array{Quantity{Int64, 𝐋, Unitful.FreeUnits{(nm,), 𝐋, nothing}}, 3}:
[:, :, 1] =
 -50 nm  -50 nm  -50 nm
 -50 nm  -50 nm  -50 nm
 -50 nm  -50 nm  -50 nm

[:, :, 2] =
 0 nm  0 nm  0 nm
 0 nm  0 nm  0 nm
 0 nm  0 nm  0 nm

[:, :, 3] =
 50 nm  50 nm  50 nm
 50 nm  50 nm  50 nm
 50 nm  50 nm  50 nm

TransferFunctions.posaxes — Function

posaxes(a::SampledArray)

Returns the positions of the axes of samples in a.

See also posgrid

julia> sa = SampledArray(reshape(1:16, (4,4)), 61u"nm");

julia> TF.posaxes(sa)
((61:61:244) nm, (61:61:244) nm)

julia> TF.posgrid(sa)[1]
4×4 Matrix{Quantity{Float64, 𝐋, Unitful.FreeUnits{(nm,), 𝐋, nothing}}}:
  61.0 nm   61.0 nm   61.0 nm   61.0 nm
 122.0 nm  122.0 nm  122.0 nm  122.0 nm
 183.0 nm  183.0 nm  183.0 nm  183.0 nm
 244.0 nm  244.0 nm  244.0 nm  244.0 nm

TransferFunctions.fillsize — Function

fillsize(Δ::Length, N::Integer) => PixelSize{N}

Construct a PixelSize with the same sampling Δ in every direction.

TransferFunctions.inner_axes — Function

inner_axes(A, edges)
inner_axes(A, K)

Determine the inner axes of the array with edges edges or when filtered with kernel K.

julia> TF.inner_axes(ones(100,100), ((2,4), (1,10)))
(3:96, 2:90)

julia> TF.inner_axes(ones(100,100), OAs.OffsetArray(ones(11,11), -5:5, -3:7))
(6:95, 4:93)

Type Aliases

TransferFunctions.Size — Type

Size{N}

N-dimensional size of an array. Alias for NTuple{N,Int}

TransferFunctions.Coordinate — Type

Coordinate{N,T}

N-dimensional coordinate. Alias for NTuple{N,T<:Real}

TransferFunctions.PixelSize — Type

PixelSize{N}

N-dimensional pixel size. Alias for NTuple{N,Length}

TransferFunctions.OneEdge — Type

OneEdge

Alias for Tuple{Int,Int}.

An edge is a tuple with the interpretation of (start, end) that determines the padding applied in a BorderArray or a TaperedArray.

See also Edges

TransferFunctions.Edges — Type

Edges{N}

Edges of an N-dimensional array. Alias for NTuple{N,OneEdge}

See also OneEdge, TaperedArray, BorderArray

Unit Types

TransferFunctions.Frequency — Type

TransferFunctions.Frequency{T, U}

A supertype for quantities and levels of dimension Unitful.𝐋 ^ -1 with a value of type T and units U.

See also: Unitful.Quantity, Unitful.Level.

TransferFunctions.FrequencyUnits — Type

TransferFunctions.FrequencyUnits{U}

A supertype for units of dimension Unitful.𝐋 ^ -1. Equivalent to Unitful.Units{U, Unitful.𝐋 ^ -1}.

See also: Unitful.Units.

TransferFunctions.FrequencyFreeUnits — Type

TransferFunctions.FrequencyFreeUnits{U}

A supertype for Unitful.FreeUnits of dimension Unitful.𝐋 ^ -1. Equivalent to Unitful.FreeUnits{U, Unitful.𝐋 ^ -1}.